==> 3. Analysis scripts/2. Adversity_summary_2022-08-18.Rmd <==
---
title: "Adversity-DNAm adolescence. Part 2: adversity summary"
output: html_notebook
editor_options: 
  chunk_output_type: console
---

*Setup*
```{r setup}
library(ggplot2)
library(gridExtra)
library(limma)
library(psych)
library(tableone)
library(tidyr)
library(viridis)

load("~/Dropbox (Partners HealthCare)/Dunn_shared/ALSPAC/Epigenetic/Paper Draft - Adversity->DNAm age17/Pacruelty_rerun/Results/age15.results.variable_2022-08-18.Rdata")

load("~/Dropbox (Partners HealthCare)/Dunn_shared/ALSPAC/Epigenetic/Aim 3/Adversity_prelim/covariates.f15.20200224.Rdata", verbose=T)

dat.path ="~/Dropbox (Partners HealthCare)/Dunn_shared/ALSPAC/Epigenetic/Aim 3/Adversity_prelim/Results/20200513/"

```

*Figure 1 - summary of adversity prevalences*
```{r figure 1}
adversity.summary <- do.call(rbind, lapply(levels(factor(age15.results.variable$Adversity)),function(i){
  print(i)
  if(i == "r_faminstability"){
    load("/Users/alexlussier/Dropbox (Partners HealthCare)/Dunn_shared/ALSPAC/Epigenetic/Aim 3/Adversity_prelim/Results/r_faminstability_exposure_data_2021-04-24.Rdata")
  }
  
  else{
    if(i=="Fscore"){
      load("/Users/alexlussier/Dropbox (Partners HealthCare)/Dunn_shared/ALSPAC/Epigenetic/Aim 3/Adversity_prelim/Results/data/Fscore_fixed_exposure_data_2021-06-01.Rdata")
      head(adversity.data.2)
      adversity.data.2 <- adversity.data.2[,c(8,2:7,9,10)]
      colnames(adversity.data.2) <- gsub("fixed_","", colnames(adversity.data.2))
    }
    
    else{
      if(i=="parcruelty"){
        load("/Users/alexlussier/Dropbox (Partners HealthCare)/Dunn_shared/ALSPAC/Epigenetic/Aim 3/Adversity_prelim/Results/data/parcruelty_exposure_data_2022-07-13.Rdata")
      }
      else{
    dat.files <- list.files(dat.path, full.names = T)
    dat.files <- dat.files[grep(i, dat.files)]
    dat.files <- dat.files[grep("exposure_data", dat.files)]
    load(dat.files)
      }
    }
  }
  colnames(adversity.data.2)
  adversity.data.2 <- adversity.data.2[match(covariate.data.f15$cidB1471,
                                             adversity.data.2$cidB1471),]
  adversity.data.3 <- cbind(adversity.data.2, covariate.data.f15)
  adversity.data.3 <- adversity.data.3[complete.cases(adversity.data.3),]
  dim(adversity.data.3)
  
  temp <- do.call(rbind, lapply(2:ncol(adversity.data.2), function(x){
    a <- data.frame(summary(as.factor(adversity.data.3[,x])))
    colnames(a)[1] <- "Individuals"
    a$Adversity <- i
    a$Variable_Name <- gsub(paste(i,"_", sep =""), "", colnames(adversity.data.3)[x])
    a$Exposure <- rownames(a)
    a$Number <- nrow(adversity.data.3)
    a$Percent <- summary(as.factor(adversity.data.3[,x]))/nrow(adversity.data.3)*100
    a
  }))
  temp <- temp[,c(2,3,4,5,1,6)]
  temp <- rbind(data.frame(Adversity = i, Variable_Name = "N",
                           Exposure = "", Number = nrow(adversity.data.3),
                           Individuals = nrow(adversity.data.3), Percent = 100),
                temp)
  
  temp
}))
adversity.summary
rownames(adversity.summary) <- 1:nrow(adversity.summary)
head(adversity.summary)
adversity.summary$Variable_Name <- gsub("D","",adversity.summary$Variable_Name)
adversity.summary$Variable_Name <- factor(adversity.summary$Variable_Name, 
                                          levels = c("N","8m","18m","21m","30m","33m","42m",
                                                     "47m","57m","61m","69m","6y","81m",
                                                     "7y","8y","9y","10y","11y",
                                                     "ever","accumulation"))

adversity.summary$Timing <- "Ever"
adversity.summary$Timing[which(adversity.summary$Variable_Name == "N")] <- "N"
adversity.summary$Timing[which(adversity.summary$Variable_Name %in% c("8m","18m","21m","30m","33m"))] <- "very_early"
adversity.summary$Timing[which(adversity.summary$Variable_Name %in% c("42m","47m","57m","61m"))] <- "early"
adversity.summary$Timing[which(adversity.summary$Variable_Name %in% c("69m","6y","81m","7y"))] <- "middle"
adversity.summary$Timing[which(adversity.summary$Variable_Name %in% c("8y","9y","10y","11y"))] <- "late"
adversity.summary$Timing <- factor(adversity.summary$Timing, levels = c("N","very_early","early",
                                                                        "middle","late","Ever"))

adv.labels <- c("Caregiver physical or emotional abuse", 
                "Sexual or physical abuse (by anyone)", 
                "Maternal psychopathology", 
                "One adult in the household", 
                "Family instability", 
                "Financial hardship", 
                "Neighborhood disadvantage")
adversity.summary$Adversity.2 <- factor(adversity.summary$Adversity, 
                                           levels = c("parcruelty", "abuse", "mompsych",
                                                      "oneadult","r_faminstability", 
                                                      "Fscore","nbhqual"))
  
library(viridis)
a <- ggplot(adversity.summary[which(adversity.summary$Variable_Name!='accumulation' &
                                 adversity.summary$Exposure !=0 ),],
       aes(x= Variable_Name, y= Adversity.2, fill = Timing))+
  geom_tile()+
  geom_tile(data =adversity.summary[which(adversity.summary$Variable_Name!='accumulation' &
                                            adversity.summary$Exposure ==1 ),], col="black")+
  theme_classic()+
  geom_text(data = adversity.summary[which(adversity.summary$Variable_Name!='accumulation' &
                                             adversity.summary$Exposure ==1 ),],
            aes(label = paste(round(Percent,1),"%",sep="")), col = 'white', size =4.5)+
  geom_text(data = adversity.summary[which(adversity.summary$Variable_Name =='N'),],
            aes(label = Individuals), col = 'black', size =6)+
  #scale_fill_viridis(discrete=T)+
  scale_fill_manual(values = c("white", viridis(n=6)[1:4], "darkgrey"))+
  xlab("Measurement timepoint (months)")+
  theme(axis.text = element_text(size =14),
        axis.title = element_text(size = 16),
        text = element_text(family = "Arial"),
        axis.title.y=element_blank(),
        legend.position = "none")+
  scale_x_discrete(position = "top", 
                   labels = c("N", c(c(8, 18, 21,30,33,42,47,57, 
                                                 61,69, 6*12, 81)/12,7, 8,9,10,11)*12,
                              "ever")) +
  scale_y_discrete(limits = rev(levels(factor(adversity.summary$Adversity.2))),
                   labels = rev(adv.labels))+
  geom_vline(xintercept = 1.5, col ='black', linetype=1)+
  geom_vline(xintercept = 6.5, col ='black', linetype=3)+
  geom_vline(xintercept = 10.5, col ='black', linetype=3)+
  geom_vline(xintercept = 14.5, col ='black', linetype=3)+
  geom_vline(xintercept = 18.5, col ='black', linetype=3)

ggsave(file = "~/Dropbox (Partners HealthCare)/Dunn_shared/ALSPAC/Epigenetic/Paper Draft - Adversity->DNAm age17/Submission 9 - Lancet Child Adolescent Health - Resubmission 2/Figures/Figure1_2023-04-17.eps", dpi = 300, width=14, height=6, device=cairo_ps)

## Accumulation distribution
ggplot(adversity.summary[which(adversity.summary$Variable_Name=='accumulation' &
                                 adversity.summary$Exposure !=0 ),],
       aes(x= Adversity.2, y = Percent, group = as.factor(Exposure), fill = as.factor(Exposure)))+
  geom_bar(stat='identity', position = position_dodge2(preserve = "single", width = 0.5))+
  theme_classic()+
  geom_text(aes(label = round(Percent,1)), 
            position = position_dodge2(preserve ="single", width = 1), hjust=-0.2)+
  scale_fill_viridis(discrete=T, "Number of exposures (accumulation)")+
  ylab("Percent of participants")+
  scale_x_discrete(labels  = adv.labels, "")+
  coord_flip()
  
ggplot(adversity.summary[which(adversity.summary$Variable_Name=='accumulation' &
                                 adversity.summary$Exposure !=0 ),],
       aes(x= Adversity.2, y = Percent, group = as.factor(Exposure), fill = as.factor(Exposure)))+
  geom_bar(stat='identity', position = position_dodge2(preserve = "single", width = 0.5))+
  theme_classic()+
  geom_text(aes(label = round(Percent,1)), 
            position = position_dodge2(preserve ="single", width = 1), hjust=-0.2)+
  scale_fill_viridis(discrete=T, "Number of exposures (accumulation)")+
  ylab("Percent of participants")+
  scale_x_discrete(labels  = adv.labels, "")+
  coord_flip()
  

temp <- adversity.summary[which(adversity.summary$Variable_Name=='accumulation' ),]
accum <- temp %>% select(Adversity.2, Exposure, Percent) %>% pivot_wider(names_from = Exposure, values_from = Percent)
accum


for(i in unique(adversity.summary$Adversity)){
  if(i==unique(adversity.summary$Adversity)[1]){accumSummary <- data.frame()}
  dat <- adversity.summary[which(adversity.summary$Variable_Name=='accumulation' ),]
  dat <- dat[dat$Adversity==i,]
  inf <- unlist(lapply(1:nrow(dat), function(x){
    rep(dat$Exposure[x], dat$Individuals[x])
    }))
  inf <- as.numeric(as.character(inf))
  accumSummary <- rbind(accumSummary, 
                        data.frame(adversity = i,
                                   mean = mean(inf), 
                                   sd = sd(inf)))
  rm(dat, inf)
  }
accumSummary
rm(adv.labels, adversity.summary, dat.path, covariate.data.f15, accum, accumSummary, temp)

```

*Figure S1; Table S3 - summary correlations between adversities*
```{r figure & table S1}
dat.path ="~/Dropbox (Partners HealthCare)/Dunn_shared/ALSPAC/Epigenetic/Aim 3/Adversity_prelim/Results/20200513/"

for(i in levels(factor(age15.results.variable$Adversity))){
  print(i)
  if(i == "r_faminstability"){
    load("/Users/alexlussier/Dropbox (Partners HealthCare)/Dunn_shared/ALSPAC/Epigenetic/Aim 3/Adversity_prelim/Results/r_faminstability_exposure_data_2021-04-24.Rdata")
  }
  else{
    if(i == "Fscore"){
      load("/Users/alexlussier/Dropbox (Partners HealthCare)/Dunn_shared/ALSPAC/Epigenetic/Aim 3/Adversity_prelim/Results/data/Fscore_fixed_exposure_data_2021-06-01.Rdata")
      head(adversity.data.2)
      adversity.data.2 <- adversity.data.2[,c(8,2:7,9,10)]
      colnames(adversity.data.2) <- gsub("fixed_","", colnames(adversity.data.2))
    }
    else{
      if(i=="parcruelty"){
        load("/Users/alexlussier/Dropbox (Partners HealthCare)/Dunn_shared/ALSPAC/Epigenetic/Aim 3/Adversity_prelim/Results/data/parcruelty_exposure_data_2022-07-13.Rdata")
      }
      else{
  dat.files <- list.files(dat.path, full.names = T)
  dat.files <- dat.files[grep(i, dat.files)]
  dat.files <- dat.files[grep("exposure_data", dat.files)]
  load(dat.files)
      }
    }
  }
  if(i == levels(factor(age15.results.variable$Adversity))[1]){
    df.adversity <- data.frame(cidB1471 = adversity.data.2$cidB1471)}
  
  if(identical(as.numeric(adversity.data.2$cidB1471), 
               as.numeric(df.adversity$cidB1471)) == F){print("error")}
  
  adversity.data.2 <- adversity.data.2[,grep(i, colnames(adversity.data.2))]
  df.adversity <- cbind(df.adversity, adversity.data.2)
  
}


parc <- grep("^parcruelty", names(df.adversity), value=T)
onead <- grep("^oneadult", names(df.adversity), value=T)
abus <- grep("^abuse", names(df.adversity), value=T)
faminst <- grep("r_faminstability", names(df.adversity), value=T)
Fscor <- grep("^Fscore", names(df.adversity), value=T)
mompsy <- grep("^mompsych", names(df.adversity), value=T)
nbhq <- grep("^nbhqual", names(df.adversity), value=T)

advers <- c(parc, abus, mompsy, onead, faminst, Fscor, nbhq)
# make adv numeric:
df.adversity[,advers] <- lapply(df.adversity[,advers], function(x) as.numeric(as.character(x)))

advers.names <- c("abus", "faminst", "Fscor", "mompsy",  "nbhq","onead","parc")

# create function to return desired output:
func <- function(x){
  
  # 1. % exposed for each adv
  d <- df.adversity[, get(x)]
  d$exp <- rowSums(d, na.rm = T)
  d$exp.flag <- ifelse(d$exp == 0, 0, 1)
  r <- table(d$exp.flag, useNA = "ifany")
  
  # 2. avg correlation within each adv, across timepoints
  # get tetrachoric:
  tet <- tetrachoric(d[, 1:length(get(x))], correct=.5)
  tet.half <- tet$rho[upper.tri(tet$rho, diag = FALSE)]
  s <- mean(tet.half) # how to take mean of just upper part of matrix?
  
  u <- c(unname(r[2]), s)
}

res12 <- matrix(unlist(lapply(advers.names, func)), byrow=T, nrow=length(advers.names))
rm(func)

# 3. avg correlation with other adv (any exp):
get.exp.flag <- function(x){
  j <- ifelse(rowSums(df.adversity[, get(x)], na.rm=T) == 0, 0, 1)
}

res3 <- matrix(unlist(lapply(advers.names, get.exp.flag)), byrow=F, ncol=length(advers.names))
names(res3) <- paste0(advers.names, ".exp.flag")
colSums(res3)
#[1]  146 154 236 336 156 173 175
## matches above

d <- cor(res3)
k <- apply(d, 2, function(x) mean(x[-which.max(x)])) # mean corr values without 1.0
k
#[1] 0.09025610 -0.03497542  0.15286392  0.16086517  0.11201796  0.12659392  0.13735458
rm(get.exp.flag, res3, d)

# combine and organize:
res123 <- cbind.data.frame(adversity = advers.names, res12, k)
names(res123)[2:4] <- c("any.exposed", "avg.timepoint.tet.corr", "avg.corr.other.adv")
rm(res12, k)
res123$any.exposed.perc <- (res123$any.exposed / nrow(df.adversity)) * 100
res123 <- res123[, c(1:2, 5, 3:4)]

adv.labels <- c("Caregiver physical or emotional abuse", 
                "Sexual or physical abuse (by anyone)", 
                "Maternal psychopathology", 
                "One adult in the household", 
                "Family instability", 
                "Financial hardship", 
                "Neighborhood disadvantage")

r <- gather(res123, dataset, value, 2:5)
r <- r[r$dataset != "any.exposed", ]
r$adversity <- factor(r$adversity, 
                      levels = c("parc","abus", "mompsy", "onead", "faminst", 
                                 "Fscor", "nbhq"))
flevels <- rev(levels(r$adversity))

r.new <- r
fsize <- 14
g1 <- ggplot(r.new[r.new$dataset == "any.exposed.perc", ], 
             aes(y=value, x= adversity, fill=adversity)) + 
  geom_bar(stat='identity') + 
  scale_x_discrete("Adversity", limits=flevels, labels=rev(adv.labels)) + 
  scale_fill_viridis(option='plasma', end=0.9, discrete=T)+
  #scale_fill_discrete("Adversity", labels=adv.labels, guide=F) + 
  scale_y_continuous(breaks=c(0, 10, 20, 30, 40, 50))+ 
  #                   labels=c("", "", "20", "", "40")) + 
  ylab("Total exposed\n(%)") + 
  coord_flip() + 
  theme_classic()+
  theme(legend.position="none",
        axis.title.x = element_text(size=fsize), 
        axis.text.x = element_text(size=fsize, color = "black"),
        axis.title.y = element_text(size=fsize),
        text = element_text(family = "Arial"),
        axis.text.y = element_text(size=fsize)) + 
  theme(panel.grid.major.x = element_line())+
  #ylim(0,40)+
  ggtitle("A")
head(r.new)

g2 <- ggplot(r.new[r.new$dataset == "avg.timepoint.tet.corr", ], 
             aes(y=value, x=adversity, fill=adversity)) + 
  geom_bar(stat='identity') + 
  scale_fill_viridis(option='plasma', end=0.9, discrete=T)+
  scale_x_discrete("Adversity", limits=flevels, labels=rev(adv.labels)) + 
  scale_y_continuous(limits=c(0, 0.8), breaks=c(0, 0.2, 0.4, 0.6, 0.8, 1))+
  ylab("Correlation\nover time") + 
  coord_flip() + 
  theme_classic()+ 
  theme(legend.position="none", 
        axis.title.y=element_blank(), 
        axis.text.y=element_blank(),
        axis.title.x = element_text(size=fsize), 
        text = element_text(family = "Arial"),
        axis.text.x = element_text(size=fsize)) + 
  theme(text = element_text(size = fsize, color = "black"),
        panel.grid.major.x = element_line())+
  #ylim(0,1)+
  ggtitle("B")

g3 <- ggplot(r.new[r.new$dataset == "avg.corr.other.adv", ], 
             aes(y=abs(value), x=adversity, fill=adversity)) + 
  geom_bar(stat='identity') + 
  scale_x_discrete("Adversity", limits=flevels, labels=rev(adv.labels)) + 
  #scale_y_continuous(limits=c(0.8), breaks=c(0, 0.2, 0.4, 0.6, 0.8, 1))+
  ylab("Correlation with\nother adversities") + 
  scale_fill_viridis(option='plasma', end=0.9, discrete=T)+
  coord_flip() +
  theme_classic()+ 
  theme(legend.position="none", 
        axis.title.y=element_blank(), 
        axis.text.y=element_blank(),
        text = element_text(family = "Arial"),
        axis.title.x = element_text(size=fsize), 
        axis.text.x = element_text(size=fsize)) + 
  theme(text = element_text(size = fsize, color = "black"),
        panel.grid.major.x = element_line())+
  ylim(0, 0.2) + 
  ggtitle("C")

g <- grid.arrange(g1, g2, g3, layout_matrix = rbind(c(1,1,1,1,2,2,3,3), c(1,1,1,1,2,2,3,3)))

write.table(r.new, file = "~/Dropbox (Partners HealthCare)/Dunn_shared/ALSPAC/Epigenetic/Paper Draft - Adversity->DNAm age17/Pacruelty_rerun/Tables/Supplemental/TableS3_adversity.summary.2022-08-18.txt", 
            sep = "\t", quote=F, row.names=F, col.names=T)

rm(fsize, g, g1, g2, g3, i, mompsy, nbhq, onead, parc, r, r.new, Fscor, flevels, faminst,
   res123, dat.path, dat.files, df.adversity, advers.names, adv.labels, advers, abus,
   adversity.data.2, age15.results.variable)
```



*Table S2*
```{r}
load("~/Dropbox (Partners HealthCare)/Dunn_shared/4- Master Files/ALSPAC_beast20220124.Rdata",
     verbose=T)

#colnames(beast)[grep("parcruelty", colnames(beast))]

load("~/Dropbox (Partners HealthCare)/Dunn_shared/ALSPAC/Epigenetic/Aim 3/Adversity_prelim/beast.adversity.aim3.2020-02-20.Rdata", verbose=T)
colnames(beast.adversity.aim3)[-which(colnames(beast.adversity.aim3)%in% colnames(beast))]

## adding missing variables to beast
load("~/Dropbox (Partners HealthCare)/Dunn_shared/ALSPAC/Epigenetic/Aim 3/Adversity_prelim/recoded_faminstability_20210424.Rdata", verbose=T)
colnames(faminst.recoded.2)[-which(colnames(faminst.recoded.2)%in% colnames(beast))]

load("~/Dropbox (Partners HealthCare)/Dunn_shared/ALSPAC/Epigenetic/Aim 3/parcrueltyRevised_2022-07-13.Rdata", verbose=T)
colnames(parcrueltyRevised)[-which(colnames(parcrueltyRevised)%in% colnames(beast))]

  
load("~/Dropbox (Partners HealthCare)/Dunn_shared/ALSPAC/Epigenetic/Aim 3/Adversity_prelim/ALSPACnotwins.2018.08.27.Rdata", verbose=T)
identical(beast$cidB1471, df$cidB1471) #F

df <- df[match(paste0(beast$cidB1471, beast$qlet), paste0(df$cidB1471, df$qlet)),]
colnames(df)[-which(colnames(df)%in% colnames(beast))]


identical(beast$cidB1471, beast.adversity.aim3$cidB1471) #TRUE
identical(beast$cidB1471, faminst.recoded.2$cidB1471) #TRUE

beast <- cbind(beast, 
               beast.adversity.aim3[,-which(colnames(beast.adversity.aim3)%in% colnames(beast))],
               r_faminstability_8y = faminst.recoded.2[,-which(colnames(faminst.recoded.2)%in%
                                                                 colnames(beast))],
               parcruelty_11y = df[,"parcruelty_11y"],
               Fscore_11y= df[,"Fscore_11y"])

identical(paste0(beast$cidB1471, beast$qlet), paste0(parcrueltyRevised$cidB1471, parcrueltyRevised$qlet))
beast$parcruelty_9y <- parcrueltyRevised$parcruelty_9y
beast$parcruelty_11y <- parcrueltyRevised$parcruelty_11y

rm(beast.adversity.aim3, faminst.recoded.2,df)

load("~/Dropbox (Partners HealthCare)/Dunn_shared/ALSPAC/Epigenetic/Fscore coding issue/updated_fscore_2021-06-01.Rdata", verbose=T)
colnames(beast)[grep("Fscore_fixed",colnames(beast))]
identical(fscore_fixed$ID, paste0(beast$cidB1471, beast$qlet))#T
beast$Fscore_fixed_11y <- fscore_fixed$Fscore_fixed_11y
rm(fscore_fixed)

load("/Users/alexlussier/Dropbox (Partners HealthCare)/Dunn_shared/ALSPAC/Epigenetic/Aim 3/DNAm_samples_cells_15up_notwins.Rdata", verbose=T)

#ARIES sample 
aries.beast <- beast[paste0(beast$cidB1471, beast$qlet) %in% 
                       paste0(samples.f15.notwins$ALN, samples.f15.notwins$QLET),]
dim(aries.beast) #966


#exposed to any adversity sample 
adversities <- colnames(beast)[c(grep("parcruelty", colnames(beast)),
                 grep("^abuse", colnames(beast)),
                 grep("r_faminst", colnames(beast)),
                 grep("mompsych", colnames(beast)),
                 grep("Fscore_fixed", colnames(beast)),
                 grep("oneadult", colnames(beast)),
                 intersect(grep("D", colnames(beast)), grep("nbhqual", colnames(beast)))
                 )]
adversities <- adversities[-grep("wk", adversities)]
adversities <- adversities[-grep("sum", adversities)]
adversities
temp <- aries.beast[,adversities]

length(which(rowSums(temp)>0)) #300
length(which(rowSums(temp, na.rm = T)>0)) #647

aries.beast.exposed <- aries.beast[which(rowSums(temp, na.rm = T)>0),]
dim(aries.beast.exposed) #688

#table one 
covars <- c( "Female", "WHITE","ed_momgest", "mom_birthage","sustained.smoke", "ppregnum",
                     "birthweight.cat")

#catVars <- c("WHITE","Female","ed_momgest", "mom_birthage", "ppregnum", "sustained.smoke")

tab.alspac <- CreateCatTable(vars = covars, data = beast[beast$kz011b ==1,])
print.alspac <- as.data.frame(print(tab.alspac, printToggle = FALSE, showAllLevels = TRUE))
dim(print.alspac)

tab.aries <- CreateCatTable(vars = covars, data = aries.beast)
print.aries <- as.data.frame(print(tab.aries, printToggle = FALSE, showAllLevels = TRUE))
dim(print.aries)

tab.exposed <- CreateCatTable(vars = covars, data = aries.beast.exposed)
print.exposed <- as.data.frame(print(tab.exposed, printToggle = FALSE, showAllLevels = TRUE))
dim(print.exposed)

table <- cbind(print.alspac,  
      ARIES= print.aries$Overall, Exposed = print.exposed$Overall)
write.table(table, file="~/Dropbox (Partners HealthCare)/Dunn_shared/ALSPAC/Epigenetic/Paper Draft - Adversity->DNAm age17/Submission 8 - Lancet Child Adolescent Health - Resubmission/Results/Table_s1_covariates_summary_2022-12-02.text",
            sep="\t", quote=F, col.names=T, row.names=T)

##chi-sq tests ##
table.2 <- data.frame(table$level, 
                      ALSPAC = as.numeric(strsplit2(print.alspac[,2], "\\(")[,1]),
                      ARIES = as.numeric(strsplit2(print.aries[,2], "\\(")[,1]),
                      exposed = as.numeric(strsplit2(print.exposed[,2], "\\(")[,1]))

#table.3 <- data.frame(table$level, 
#                      ALSPAC = as.numeric(gsub("\\)","", strsplit2(print.alspac[,2], "\\(")[,2])),
#                      ARIES = as.numeric(gsub("\\)","", strsplit2(print.aries[,2], "\\(")[,2])),
#                      exposed = as.numeric(gsub("\\)","", strsplit2(print.exposed[,2], "\\(")[,2])))

#female
chisq.test(table.2[2:3,2:3]) #p =0.068
chisq.test(table.2[2:3,c(2,4)]) #p =0.113
chisq.test(table.2[2:3,c(3:4)]) #p =0.986

#white
chisq.test(table.2[4:5,2:3]) #p =0.007 
chisq.test(table.2[4:5,c(2,4)]) #p =0.38 
chisq.test(table.2[4:5,c(3:4)]) #p =0.28

#mom education
chisq.test(table.2[6:9,2:3]) #p <0.001
chisq.test(table.2[6:9,c(2,4)]) #p <0.001
chisq.test(table.2[6:9,c(3:4)]) #p= 0.49

#mom birthage
chisq.test(table.2[10:12,2:3]) #p <0.001
chisq.test(table.2[10:12,c(2,4)]) #p <0.001
chisq.test(table.2[10:12,c(3:4)]) #p =0.68

#smoking
chisq.test(table.2[13:14,2:3]) #p <0.001
chisq.test(table.2[13:14,c(2,4)]) #p <0.001
chisq.test(table.2[13:14,c(3:4)]) #p =0.192

#parity
chisq.test(table.2[15:18,2:3]) #p =0.004
chisq.test(table.2[15:18,c(2,4)]) #p =0.103 
chisq.test(table.2[15:18,c(3:4)]) #p =0.96

#birthweight
chisq.test(table.2[19:22,2:3]) #p <0.001
chisq.test(table.2[19:22,c(2,4)]) #p <0.001 
chisq.test(table.2[19:22,c(3:4)]) #p =0.98

rm(adversities,aries.beast, aries.beast.exposed,
   beast, betas.f15.win.notwins, cell.f15.notwins, covars, print.alspac, print.aries,
   print.exposed, samples.f15.notwins, tab.alspac, tab.aries, tab.exposed, table, table.2, temp,
   parcrueltyRevised)
```
