==> 3. Analysis scripts/4. Top_hits_stability_2022-12-05.Rmd <==
---
title: "Adversity-DNAm adolescence. Part 4: Stability"
output: html_document
editor_options: 
  chunk_output_type: console
---

*Setup*
```{r}
library(broom)
library(dplyr)

setwd("~/Dropbox (Partners HealthCare)/Dunn_shared/ALSPAC/Epigenetic/Paper Draft - Adversity->DNAm age17/Pacruelty_rerun/")
load("Results/age15.results.variable_2022-08-18.Rdata", verbose=T)
load("Results/age15.results.top_2022-08-24.Rdata", verbose=T)


```

*Testing associations from age 15 at age 0 and 7 and 15*
```{r}
#cell type correction for cord samples
load("/Users/alexlussier/Dropbox (Partners HealthCare)/Dunn_shared/ALSPAC_DNAm_differences/data/betas_Cord_WIN_20200128.Rdata", verbose=T)
cord.samples <- read.table("/Users/alexlussier/Dropbox (Partners HealthCare)/Dunn_shared/ALSPAC_DNAm_differences/data/samplesheet_Cord_20200121.txt", sep ="\t", header=T)
cord.cells <- read.table("/Users/alexlussier/Dropbox (Partners HealthCare)/Dunn_shared/ALSPAC_DNAm_differences/data/celltypes_Cord_20200121.txt", sep ="\t", header=T)
identical(as.character(cord.samples$Sample_Name), rownames(betas.cord.win))#TRUE
identical(as.character(cord.samples$Sample_Name), as.character(cord.cells$IID))#TRUE

cord.betas <- betas.cord.win[,colnames(betas.cord.win) %in% age15.fdr.R2$CpG]
dim(cord.betas)
rm(betas.cord.win)

cord.beta.cor <- do.call(cbind, lapply(1:ncol(cord.betas), function(i){
  print(i)
  a <- lm(cord.betas[,i] ~ Bcell+ CD4T+ CD8T+ Gran+ Mono+ NK+ nRBC, data=cord.cells )
  resid(a) + mean(cord.betas[,i])
  }))
colnames(cord.beta.cor) <- colnames(cord.betas)
rownames(cord.beta.cor) <- rownames(cord.betas)
rm(cord.betas, cord.cells)
##cord.betas.cor

#### cell type correction for age 7 ####
load("/Users/alexlussier/Dropbox (Partners HealthCare)/Dunn_shared/ALSPAC_DNAm_differences/data/betas_F7_WIN_20200128.Rdata", verbose=T)
f7.samples <- read.table("/Users/alexlussier/Dropbox (Partners HealthCare)/Dunn_shared/ALSPAC_DNAm_differences/data/samplesheet_F7_20200121.txt", sep ="\t", header=T)
f7.cells <- read.table("/Users/alexlussier/Dropbox (Partners HealthCare)/Dunn_shared/ALSPAC_DNAm_differences/data/celltypes_F7_20200121.txt", sep ="\t", header=T)
identical(as.character(f7.samples$Sample_Name), rownames(betas.f7.win))#TRUE
identical(as.character(f7.samples$Sample_Name), as.character(f7.cells$Sample_Name))#TRUE

f7.betas <- betas.f7.win[,colnames(betas.f7.win) %in% age15.fdr.R2$CpG]
dim(f7.betas)
rm(betas.f7.win)

f7.beta.cor <- do.call(cbind, lapply(1:ncol(f7.betas), function(i){
  print(i)
  a <- lm(f7.betas[,i] ~ Bcell+ CD4T+ CD8T+ Gran+ Mono+ NK, data=f7.cells )
  resid(a) + mean(f7.betas[,i])
  }))
colnames(f7.beta.cor) <- colnames(f7.betas)
rownames(f7.beta.cor) <- rownames(f7.betas)
scatter.smooth(f7.beta.cor~f7.beta.cor)
rm(f7.betas, f7.cells)
# f7.betas.cor


#### ASSOCIATION TESTING ####
covars <- c("ed_momgest", "WHITE", "Female", "mom_birthage", "ppregnum",
                     "birthweight","sustained.smoke")
                     #"sample_type")
load("~/Dropbox (Partners HealthCare)/Dunn_shared/4- Master Files/ALSPAC_beast20220124.Rdata", verbose=T)


## adding missing variables to beast 
load("~/Dropbox (Partners HealthCare)/Dunn_shared/ALSPAC/Epigenetic/Aim 3/Adversity_prelim/beast.adversity.aim3.2020-02-20.Rdata", verbose=T)
load("~/Dropbox (Partners HealthCare)/Dunn_shared/ALSPAC/Epigenetic/Aim 3/Adversity_prelim/recoded_faminstability_20210424.Rdata", verbose=T)
load("~/Dropbox (Partners HealthCare)/Dunn_shared/ALSPAC/Epigenetic/Aim 3/parcrueltyRevised_2022-07-13.Rdata", verbose=T)
load("~/Dropbox (Partners HealthCare)/Dunn_shared/ALSPAC/Epigenetic/Aim 3/Adversity_prelim/ALSPACnotwins.2018.08.27.Rdata", verbose=T)
identical(beast$cidB1471, df$cidB1471) #F

df <- df[match(paste0(beast$cidB1471, beast$qlet), paste0(df$cidB1471, df$qlet)),]
colnames(df)[-which(colnames(df)%in% colnames(beast))]
identical(beast$cidB1471, beast.adversity.aim3$cidB1471) #TRUE
identical(beast$cidB1471, faminst.recoded.2$cidB1471) #TRUE
beast <- cbind(beast, 
               beast.adversity.aim3[,-which(colnames(beast.adversity.aim3)%in% colnames(beast))],
               r_faminstability_8y = faminst.recoded.2[,-which(colnames(faminst.recoded.2)%in%
                                                                 colnames(beast))],
               parcruelty_11y = df[,"parcruelty_11y"],
               Fscore_11y= df[,"Fscore_11y"])
identical(paste0(beast$cidB1471, beast$qlet), 
          paste0(parcrueltyRevised$cidB1471, parcrueltyRevised$qlet))
beast$parcruelty_9y <- parcrueltyRevised$parcruelty_9y
beast$parcruelty_11y <- parcrueltyRevised$parcruelty_11y

#adding Fscore fixed
load("~/Dropbox (Partners HealthCare)/Dunn_shared/ALSPAC/Epigenetic/Fscore coding issue/updated_fscore_2021-06-01.Rdata", verbose=T)
colnames(beast)[grep("Fscore_",colnames(beast))]
identical(fscore_fixed$ID, paste0(beast$cidB1471, beast$qlet))#T
beast$Fscore_fixed_11y <- fscore_fixed$Fscore_fixed_11y
#cleanup fscore
remove <- colnames(beast)[grep("Fscore_",colnames(beast))]
remove <- remove[-grep("_fixed",remove)]
beast <- beast[,!colnames(beast) %in% remove]
colnames(beast) <- gsub("Fscore_fixed_","Fscore_", colnames(beast))

rm(beast.adversity.aim3, faminst.recoded.2, df, fscore_fixed, remove, parcrueltyRevised)


#Cord associations#
for(i in 1:nrow(age15.fdr.R2)){
  print(i)
  if(i == 1){cord.associations <- data.frame()}
  temp <- age15.fdr.R2[i,]
  age0 <- cord.beta.cor[,colnames(cord.beta.cor) == temp$CpG]
  
  b <- colnames(beast)[grep(paste0("^", temp$Adversity), colnames(beast))]
  if(temp$Adversity =="mompsych"){b <- b[-grep("wk",b)]}
  if(temp$Adversity =="oneadult"){b <- b[-grep("sum", b)]}
  recency.vec <- gsub("m", "", gsub(paste0(temp$Adversity,"_"),"",b))
  recency.vec[grep("y",recency.vec)] <- as.numeric(gsub("y","",
                                                          recency.vec[grep("y",recency.vec)]))*12
  recency.vec <- as.numeric(recency.vec)/12
  
  exposure <- beast[,colnames(beast) %in% c("cidB1471", covars, b)]
  exposure <- exposure[match(cord.samples$ALN, exposure$cidB1471),]
  age0 <- age0[complete.cases(exposure)]
  exposure <- exposure[complete.cases(exposure),]
  
  if(temp$Variable_Name == "recency"){
    exp <- do.call("cbind", lapply(exposure[, colnames(exposure) %in% b], 
                                   function(x) as.numeric(as.character(x))))
    exposure <- exposure %>% 
      mutate(accumulation = rowSums(exp)) %>% 
      mutate(ever = ifelse(accumulation > 0, 1, 0))
    exposure$recency <-  rowSums(exp %*% diag(recency.vec))
    
    colnames(exposure)[colnames(exposure)=="recency"] <- "exposure"
    rm(exp)
  }
  if(temp$Variable_Name == "accumulation"){
    exp <- do.call("cbind", lapply(exposure[, colnames(exposure) %in% b], 
                                   function(x) as.numeric(as.character(x))))
    exposure <- exposure %>% 
      mutate(accumulation = rowSums(exp)) %>% 
      mutate(ever = ifelse(accumulation > 0, 1, 0))
    exposure$recency <-  rowSums(exp %*% diag(recency.vec))
    
    colnames(exposure)[colnames(exposure)=="accumulation"] <- "exposure"
    rm(exp)
  }
  
  else{
  colnames(exposure)[colnames(exposure) == temp$Variable_Name] <- "exposure"
  }
  exposure$sum <- rowSums(exposure[,colnames(exposure) %in% b])
  
  mod <- lm(age0 ~ exposure + ed_momgest +WHITE + Female + 
                  mom_birthage + ppregnum + birthweight +
                  sustained.smoke, data = exposure)
  
  res <-tidy(mod)
  conf <- confint(mod)
  
  a <- data.frame(Cpg = temp$CpG,
             Variable_Name = temp$Variable_Name, 
             Model = temp$Model,
             Timing = temp$Timing,
             Adversity = temp$Adversity, 
             N = nrow(exposure),
             p.value = res$p.value[2],
             estimate = res$estimate[2],
             statistic = res$statistic[2],
             std.error = res$std.error[2],
             CI_lower = conf[2,1],
             CI_high = conf[2,2],
             beta.unexp = mean(age0[which(exposure$exposure==0)]),
             beta.exp = mean(age0[which(exposure$exposure>0)]),
             beta.unexp.all = mean(age0[which(exposure$sum==0 & exposure$exposure==0)]),
             beta.exp.other = mean(age0[which(exposure$sum>0)]),
             shapiro = shapiro.test(rstandard(mod))$p.value)
  a$delta <- a$beta.exp - a$beta.unexp
  
  cord.associations <- rbind(cord.associations, a)
  
  rm(res, a,b, exposure, age0, temp, recency.vec)
}

head(cord.associations)
cord.associations$FDR <- p.adjust(cord.associations$p.value, method = "BH")
cord.associations$bonf <- p.adjust(cord.associations$p.value, method = "bonferroni")
sum(cord.associations$p.value<0.05) #4
sum(cord.associations$FDR<0.05) #0
sum(cord.associations$bonf<0.05) #0
max(cord.associations$shapiro)

save(cord.associations, file = "~/Dropbox (Partners HealthCare)/Dunn_shared/ALSPAC/Epigenetic/Paper Draft - Adversity->DNAm age17/Pacruelty_rerun/Results/Stability/cord.associations.age15.2022-12-07.Rdata")
write.table(cord.associations, 
     file = "~/Dropbox (Partners HealthCare)/Dunn_shared/ALSPAC/Epigenetic/Paper Draft - Adversity->DNAm age17/Pacruelty_rerun/Tables/Stability/cord.associations.age15.hits.2022-12-07.txt", sep ="\t", 
            col.names = T, row.names=F, quote=F)


# Age 7 associations #
for(i in 1:nrow(age15.fdr.R2)){
  print(i)
  if(i == 1){f7.associations <- data.frame()}
  temp <- age15.fdr.R2[i,]
  age7 <- f7.beta.cor[,colnames(f7.beta.cor) == temp$CpG]
  

  b <- colnames(beast)[grep(paste0("^", temp$Adversity), colnames(beast))]
  if(temp$Adversity =="mompsych"){b <- b[-grep("wk",b)]}
  if(temp$Adversity =="oneadult"){b <- b[-grep("sum", b)]}
  recency.vec <- gsub("m", "", gsub(paste0(temp$Adversity,"_"),"",b))
  recency.vec[grep("y",recency.vec)] <- as.numeric(gsub("y","",
                                                          recency.vec[grep("y",recency.vec)]))*12
  recency.vec <- as.numeric(recency.vec)/12

  
  exposure <- beast[,colnames(beast) %in% c("cidB1471", covars, b)]
  exposure <- exposure[match(f7.samples$ALN, exposure$cidB1471),]
  age7 <- age7[complete.cases(exposure)]
  exposure <- exposure[complete.cases(exposure),]
  
  if(temp$Variable_Name == "recency"){
    exp <- do.call("cbind", lapply(exposure[, colnames(exposure) %in% b], 
                                   function(x) as.numeric(as.character(x))))
    exposure <- exposure %>% 
      mutate(accumulation = rowSums(exp)) %>% 
      mutate(ever = ifelse(accumulation > 0, 1, 0))
    exposure$recency <-  rowSums(exp %*% diag(recency.vec))
    
    colnames(exposure)[colnames(exposure)=="recency"] <- "exposure"
    rm(exp)
  }
  
  if(temp$Variable_Name == "accumulation"){
    exp <- do.call("cbind", lapply(exposure[, colnames(exposure) %in% b], 
                                   function(x) as.numeric(as.character(x))))
    exposure <- exposure %>% 
      mutate(accumulation = rowSums(exp)) %>% 
      mutate(ever = ifelse(accumulation > 0, 1, 0))
    exposure$recency <-  rowSums(exp %*% diag(recency.vec))
    
    colnames(exposure)[colnames(exposure)=="accumulation"] <- "exposure"
    rm(exp)
  }
  
  else{
  colnames(exposure)[colnames(exposure) == temp$Variable_Name] <- "exposure"
  }
  exposure$sum <- rowSums(exposure[,colnames(exposure) %in% b])
  mod <- lm(age7 ~ exposure + ed_momgest +WHITE + Female + 
                  mom_birthage + ppregnum + birthweight +
                  sustained.smoke, data = exposure)
  res <-tidy(mod)
  conf <- confint(mod)
  
  a <- data.frame(Cpg = temp$CpG,
             Variable_Name = temp$Variable_Name, 
             Model = temp$Model,
             Timing = temp$Timing,
             Adversity = temp$Adversity, 
             p.value = res$p.value[2],
             estimate = res$estimate[2],
             statistic = res$statistic[2],
             std.error = res$std.error[2],
             CI_lower = conf[2,1],
             CI_high = conf[2,2],
             beta.unexp = mean(age7[which(exposure$exposure==0)]),
             beta.exp = mean(age7[which(exposure$exposure>0)]),
             beta.unexp.all = mean(age7[which(exposure$sum==0 & exposure$exposure==0)]),
             beta.exp.other = mean(age7[which(exposure$sum>0)]),
             shapiro = shapiro.test(rstandard(mod))$p.value)
  a$delta <- a$beta.exp - a$beta.unexp
  
  f7.associations <- rbind(f7.associations, a)
  
  rm(res, a, b, exposure, age7, temp, recency.vec)
}

head(f7.associations)
f7.associations$FDR <- p.adjust(f7.associations$p.value, method = "BH")
f7.associations$bonf <- p.adjust(f7.associations$p.value, method = "bonferroni")
sum(f7.associations$p.value<0.05) #5
sum(f7.associations$FDR<0.05) #0
sum(f7.associations$bonf<0.05) #0
max(f7.associations$shapiro) #lmao

save(f7.associations, file = "~/Dropbox (Partners HealthCare)/Dunn_shared/ALSPAC/Epigenetic/Paper Draft - Adversity->DNAm age17/Pacruelty_rerun/Results/Stability/age7.associations.age15.2022-12-07.Rdata")

write.table(f7.associations, file = "~/Dropbox (Partners HealthCare)/Dunn_shared/ALSPAC/Epigenetic/Paper Draft - Adversity->DNAm age17/Pacruelty_rerun/Tables/Stability/age7.associations.age15.hits.2022-12-07.txt", sep ="\t", 
            col.names = T, row.names=F, quote=F)

#### cell type correction for age 15 ####
load("/Users/alexlussier/Dropbox (Partners HealthCare)/Dunn_shared/ALSPAC/Epigenetic/Aim 3/DNAm_samples_cells_15up_notwins.Rdata", verbose=T)

identical(as.character(samples.f15.notwins$Sample_Name), rownames(betas.f15.win.notwins))#TRUE
identical(as.character(samples.f15.notwins$Sample_Name),
          as.character(cell.f15.notwins$Sample_Name))#TRUE

f15.betas <- betas.f15.win.notwins[,colnames(betas.f15.win.notwins) %in% age15.fdr.R2$CpG]
rm(betas.f15.win.notwins)

f15.beta.cor <- do.call(cbind, lapply(1:ncol(f15.betas), function(i){
  print(i)
  a <- lm(f15.betas[,i] ~ Bcell+ CD4T+ CD8T+ Gran+ Mono+ NK, data=cell.f15.notwins )
  resid(a) + mean(f15.betas[,i])
  }))
colnames(f15.beta.cor) <- colnames(f15.betas)
rownames(f15.beta.cor) <- rownames(f15.betas)
rm(f15.betas, cell.f15.notwins)

## ASSOCIATION TESTING AGE 15##

for(i in 1:nrow(age15.fdr.R2)){
  print(i)
  if(i == 1){f15.associations <- data.frame()}
  temp <- age15.fdr.R2[i,]
  age15 <- f15.beta.cor[,colnames(f15.beta.cor) == temp$CpG]

  b <- colnames(beast)[grep(paste0("^", temp$Adversity), colnames(beast))]
  if(temp$Adversity =="mompsych"){b <- b[-grep("wk",b)]}
  if(temp$Adversity =="oneadult"){b <- b[-grep("sum", b)]}
  recency.vec <- gsub("m", "", gsub(paste0(temp$Adversity,"_"),"",b))
  recency.vec[grep("y",recency.vec)] <- as.numeric(gsub("y","",
                                                             recency.vec[grep("y",recency.vec)]))*12
  recency.vec <- as.numeric(recency.vec)/12

  
  exposure <- beast[,colnames(beast) %in% c("cidB1471", covars, b)]
  exposure <- exposure[match(samples.f15.notwins$ALN, exposure$cidB1471),]
  age15 <- age15[complete.cases(exposure)]
  exposure <- exposure[complete.cases(exposure),]
  
  if(temp$Variable_Name == "recency"){
    exp <- do.call("cbind", lapply(exposure[, colnames(exposure) %in% b], 
                                   function(x) as.numeric(as.character(x))))
    exposure <- exposure %>% 
      mutate(accumulation = rowSums(exp)) %>% 
      mutate(ever = ifelse(accumulation > 0, 1, 0))
    exposure$recency <-  rowSums(exp %*% diag(recency.vec))
    
    colnames(exposure)[colnames(exposure)=="recency"] <- "exposure"
    rm(exp)
  }
  
  if(temp$Variable_Name == "accumulation"){
    exp <- do.call("cbind", lapply(exposure[, colnames(exposure) %in% b], 
                                   function(x) as.numeric(as.character(x))))
    exposure <- exposure %>% 
      mutate(accumulation = rowSums(exp)) %>% 
      mutate(ever = ifelse(accumulation > 0, 1, 0))
    exposure$recency <-  rowSums(exp %*% diag(recency.vec))
    
    colnames(exposure)[colnames(exposure)=="accumulation"] <- "exposure"
    rm(exp)
  }
  else{
  colnames(exposure)[colnames(exposure) == temp$Variable_Name] <- "exposure"
  }
  exposure$sum <- rowSums(exposure[,colnames(exposure) %in% b])
  
  mod <- glm(age15 ~ exposure + ed_momgest +WHITE + Female + 
                  mom_birthage + ppregnum + birthweight +
                  sustained.smoke, data = exposure)
  
  res <-tidy(mod)
  conf <- confint(mod)
  shapiro.test(rstandard(mod))
  
  a <- data.frame(Cpg = temp$CpG,
             Variable_Name = temp$Variable_Name, 
             Model = temp$Model,
             Timing = temp$Timing,
             Adversity = temp$Adversity, 
             p.value = res$p.value[2],
             estimate = res$estimate[2],
             statistic = res$statistic[2],
             std.error = res$std.error[2],
             CI_lower = conf[2,1],
             CI_high = conf[2,2],
             beta.unexp = mean(age15[which(exposure$exposure==0)]),
             beta.exp = mean(age15[which(exposure$exposure>0)]),
             beta.unexp.all = mean(age15[which(exposure$sum==0 & exposure$exposure==0)]),
             beta.exp.other = mean(age15[which(exposure$sum>0)]),
             shapiro = shapiro.test(rstandard(mod))$p.value)
  a$delta <- a$beta.exp - a$beta.unexp
  
  f15.associations <- rbind(f15.associations, a)
  
  rm(res, a, b, exposure, age15, temp, recency.vec)
}
f15.associations$p.value
f15.associations$FDR <- p.adjust(f15.associations$p.value, method = "BH")
f15.associations$bonf <- p.adjust(f15.associations$p.value, method = "bonferroni")
sum(f15.associations$p.value<0.05) #41
sum(f15.associations$FDR<0.05) #41
sum(f15.associations$bonf<0.05) #41
max(f15.associations$shapiro)
sum(f15.associations$shapiro>0.05) #2

save(f15.associations, file="~/Dropbox (Partners HealthCare)/Dunn_shared/ALSPAC/Epigenetic/Paper Draft - Adversity->DNAm age17/Pacruelty_rerun/Results/Stability/age15.associations.age15.2022-12-07.Rdata")

write.table(f15.associations, file = "~/Dropbox (Partners HealthCare)/Dunn_shared/ALSPAC/Epigenetic/Paper Draft - Adversity->DNAm age17/Pacruelty_rerun/Tables/Stability/age15.associations.age15.hits.2022-12-07.txt", sep ="\t", 
            col.names = T, row.names=F, quote=F)

shapiro.test(mod$residuals)

rm(beast, covars, conf,
   f7.beta.cor, f7.samples,
   cord.beta.cor, cord.samples,
   cord.associations, f7.associations, i,
   f15.associations, f15.beta.cor, samples.f15.notwins)
```

*Testing associations from age 7 at age 0 and 7 and 15*
```{r}
load("~/Dropbox (Partners HealthCare)/Dunn_shared/ALSPAC/Epigenetic/Paper Draft - Data differences at age 7/DNAm-data-diff-analysis/1.Clean_Aug2021/Results/age7.all.results.2021-09-21.Rdata", verbose=T)

results <- age7.all.results[age7.all.results$Analysis==6,]
dim(results)/7
sum(results$FDR<0.05) #46
new.results.fdr <- results[results$FDR<0.05,]
rm(age7.all.results)

#cell type correction for cord samples
load("/Users/alexlussier/Dropbox (Partners HealthCare)/Dunn_shared/ALSPAC_DNAm_differences/data/betas_Cord_WIN_20200128.Rdata", verbose=T)
cord.samples <- read.table("/Users/alexlussier/Dropbox (Partners HealthCare)/Dunn_shared/ALSPAC_DNAm_differences/data/samplesheet_Cord_20200121.txt", sep ="\t", header=T)
cord.cells <- read.table("/Users/alexlussier/Dropbox (Partners HealthCare)/Dunn_shared/ALSPAC_DNAm_differences/data/celltypes_Cord_20200121.txt", sep ="\t", header=T)
identical(as.character(cord.samples$Sample_Name), rownames(betas.cord.win))#TRUE
identical(as.character(cord.samples$Sample_Name), as.character(cord.cells$IID))#TRUE

cord.betas <- betas.cord.win[,colnames(betas.cord.win) %in% new.results.fdr$CpG]
dim(cord.betas)
rm(betas.cord.win)

cord.beta.cor <- do.call(cbind, lapply(1:ncol(cord.betas), function(i){
  print(i)
  a <- lm(cord.betas[,i] ~ Bcell+ CD4T+ CD8T+ Gran+ Mono+ NK+ nRBC, data=cord.cells )
  resid(a) + mean(cord.betas[,i])
  }))
colnames(cord.beta.cor) <- colnames(cord.betas)
rownames(cord.beta.cor) <- rownames(cord.betas)
scatter.smooth(cord.beta.cor~cord.beta.cor)
rm(cord.betas, cord.cells)
##cord.betas.cor

#### cell type correction for age 7 ####
load("/Users/alexlussier/Dropbox (Partners HealthCare)/Dunn_shared/ALSPAC_DNAm_differences/data/betas_F7_WIN_20200128.Rdata", verbose=T)
f7.samples <- read.table("/Users/alexlussier/Dropbox (Partners HealthCare)/Dunn_shared/ALSPAC_DNAm_differences/data/samplesheet_F7_20200121.txt", sep ="\t", header=T)
f7.cells <- read.table("/Users/alexlussier/Dropbox (Partners HealthCare)/Dunn_shared/ALSPAC_DNAm_differences/data/celltypes_F7_20200121.txt", sep ="\t", header=T)
identical(as.character(f7.samples$Sample_Name), rownames(betas.f7.win))#TRUE
identical(as.character(f7.samples$Sample_Name), as.character(f7.cells$Sample_Name))#TRUE

f7.betas <- betas.f7.win[,colnames(betas.f7.win) %in% new.results.fdr$CpG]
dim(f7.betas)
rm(betas.f7.win)

f7.beta.cor <- do.call(cbind, lapply(1:ncol(f7.betas), function(i){
  print(i)
  a <- lm(f7.betas[,i] ~ Bcell+ CD4T+ CD8T+ Gran+ Mono+ NK, data=f7.cells )
  resid(a) + mean(f7.betas[,i])
  }))
colnames(f7.beta.cor) <- colnames(f7.betas)
rownames(f7.beta.cor) <- rownames(f7.betas)
scatter.smooth(f7.beta.cor~f7.beta.cor)
rm(f7.betas, f7.cells)
# f7.betas.cor

#### cell type correction for age 15 ####
load("/Users/alexlussier/Dropbox (Partners HealthCare)/Dunn_shared/ALSPAC/Epigenetic/Aim 3/DNAm_samples_cells_15up_notwins.Rdata", verbose=T)

identical(as.character(samples.f15.notwins$Sample_Name), rownames(betas.f15.win.notwins))#TRUE
identical(as.character(samples.f15.notwins$Sample_Name),
          as.character(cell.f15.notwins$Sample_Name))#TRUE

f15.betas <- betas.f15.win.notwins[,colnames(betas.f15.win.notwins) %in% new.results.fdr$CpG]
rm(betas.f15.win.notwins)

f15.beta.cor <- do.call(cbind, lapply(1:ncol(f15.betas), function(i){
  print(i)
  a <- lm(f15.betas[,i] ~ Bcell+ CD4T+ CD8T+ Gran+ Mono+ NK, data=cell.f15.notwins )
  resid(a) + mean(f15.betas[,i])
  }))
colnames(f15.beta.cor) <- colnames(f15.betas)
rownames(f15.beta.cor) <- rownames(f15.betas)
rm(f15.betas, cell.f15.notwins)

## ASSOCIATION TESTING ##
covars <- c("ed_momgest", "WHITE", "Female", "mom_birthage", "ppregnum",
                     "birthweight","sustained.smoke")
                     #"sample_type")
load("~/Dropbox (Partners HealthCare)/Dunn_shared/4- Master Files/ALSPAC_beast20220124.Rdata",
     verbose=T)

#age 0
for(i in 1:nrow(new.results.fdr)){
  print(i)
  if(i == 1){cord.associations <- data.frame()}
  temp <- new.results.fdr[i,]
  age0 <- cord.beta.cor[,colnames(cord.beta.cor) == temp$CpG]
  

  b <- colnames(beast)[grep(paste0("^", temp$Adversity), colnames(beast))]
  if(temp$Adversity =="mompsych"){b <- b[-grep("wk",b)]}
  if(temp$Adversity =="oneadult"){b <- b[-grep("sum", b)]}
  recency.vec <- gsub("m", "", gsub(paste0(temp$Adversity,"_"),"",b))
  recency.vec[grep("y",recency.vec)] <- as.numeric(gsub("y","",
                                                            recency.vec[grep("y",recency.vec)]))*12
  recency.vec <- as.numeric(recency.vec)/12

  
  exposure <- beast[,colnames(beast) %in% c("cidB1471", covars, b)]
  exposure <- exposure[match(cord.samples$ALN, exposure$cidB1471),]
  age0 <- age0[complete.cases(exposure)]
  exposure <- exposure[complete.cases(exposure),]
  
  if(temp$Variable_Name == "recency"){
    exp <- do.call("cbind", lapply(exposure[, colnames(exposure) %in% b], 
                                   function(x) as.numeric(as.character(x))))
    exposure <- exposure %>% 
      mutate(accumulation = rowSums(exp)) %>% 
      mutate(ever = ifelse(accumulation > 0, 1, 0))
    exposure$recency <-  rowSums(exp %*% diag(recency.vec))
    
    colnames(exposure)[colnames(exposure)=="recency"] <- "exposure"
    rm(exp)
  }
  
  if(temp$Variable_Name == "accumulation"){
    exp <- do.call("cbind", lapply(exposure[, colnames(exposure) %in% b], 
                                   function(x) as.numeric(as.character(x))))
    exposure <- exposure %>% 
      mutate(accumulation = rowSums(exp)) %>% 
      mutate(ever = ifelse(accumulation > 0, 1, 0))
    exposure$recency <-  rowSums(exp %*% diag(recency.vec))
    
    colnames(exposure)[colnames(exposure)=="accumulation"] <- "exposure"
    rm(exp)
  }
  
  else{
  colnames(exposure)[colnames(exposure) == temp$Variable_Name] <- "exposure"
  }
  exposure$sum <- rowSums(exposure[,colnames(exposure) %in% b])
  
  res <-tidy(lm(age0 ~ exposure + ed_momgest +WHITE + Female + 
                  mom_birthage + ppregnum + birthweight +
                  sustained.smoke, data = exposure))
  conf <- confint(lm(age0 ~ exposure + ed_momgest +WHITE + Female + 
                  mom_birthage + ppregnum + birthweight +
                  sustained.smoke, data = exposure))
  
  a <- data.frame(Cpg = temp$CpG,
             Variable_Name = temp$Variable_Name, 
             Model = temp$Model,
             Timing = temp$Timing,
             Adversity = temp$Adversity, 
             p.value = res$p.value[2],
             estimate = res$estimate[2],
             statistic = res$statistic[2],
             std.error = res$std.error[2],
             CI_lower = conf[2,1],
             CI_high =conf[2,2],
             beta.unexp = mean(age0[which(exposure$exposure==0)]),
             beta.exp = mean(age0[which(exposure$exposure>0)]),
             beta.unexp.all = mean(age0[which(exposure$sum==0 & exposure$exposure==0)]),
             beta.exp.other = mean(age0[which(exposure$sum>0)]))
  a$delta <- a$beta.exp - a$beta.unexp
  
  cord.associations <- rbind(cord.associations, a)
  
  rm(res, a, b, exposure, age0, temp, recency.vec)
}

head(cord.associations)
cord.associations$FDR <- p.adjust(cord.associations$p.value, method = "BH")
cord.associations$bonf <- p.adjust(cord.associations$p.value, method = "bonferroni")
sum(cord.associations$p.value<0.05) #3
sum(cord.associations$FDR<0.05) #0
sum(cord.associations$bonf<0.05) #0

save(cord.associations, file = "~/Dropbox (Partners HealthCare)/Dunn_shared/ALSPAC/Epigenetic/Paper Draft - Adversity->DNAm age17/Pacruelty_rerun/Results/Stability/cord.associations.age7.2022-12-05.Rdata")
write.table(cord.associations, file = "~/Dropbox (Partners HealthCare)/Dunn_shared/ALSPAC/Epigenetic/Paper Draft - Adversity->DNAm age17/Pacruelty_rerun/Tables/Stability/cord.associations.age7.2022-12-05.txt", sep ="\t", 
            col.names = T, row.names=F, quote=F)

#age 7
for(i in 1:nrow(new.results.fdr)){
  print(i)
  if(i == 1){f7.associations <- data.frame()}
  temp <- new.results.fdr[i,]
  age7 <- f7.beta.cor[,colnames(f7.beta.cor) == temp$CpG]
  

  b <- colnames(beast)[grep(paste0("^", temp$Adversity), colnames(beast))]
  if(temp$Adversity =="mompsych"){b <- b[-grep("wk",b)]}
  if(temp$Adversity =="oneadult"){b <- b[-grep("sum", b)]}
  recency.vec <- gsub("m", "", gsub(paste0(temp$Adversity,"_"),"",b))
  recency.vec[grep("y",recency.vec)] <- as.numeric(gsub("y","",
                                                            recency.vec[grep("y",recency.vec)]))*12
  recency.vec <- as.numeric(recency.vec)/12

  
  exposure <- beast[,colnames(beast) %in% c("cidB1471", covars, b)]
  exposure <- exposure[match(f7.samples$ALN, exposure$cidB1471),]
  age7 <- age7[complete.cases(exposure)]
  exposure <- exposure[complete.cases(exposure),]
  
  if(temp$Variable_Name == "recency"){
    exp <- do.call("cbind", lapply(exposure[, colnames(exposure) %in% b], 
                                   function(x) as.numeric(as.character(x))))
    exposure <- exposure %>% 
      mutate(accumulation = rowSums(exp)) %>% 
      mutate(ever = ifelse(accumulation > 0, 1, 0))
    exposure$recency <-  rowSums(exp %*% diag(recency.vec))
    
    colnames(exposure)[colnames(exposure)=="recency"] <- "exposure"
    rm(exp)
  }
  
  if(temp$Variable_Name == "accumulation"){
    exp <- do.call("cbind", lapply(exposure[, colnames(exposure) %in% b], 
                                   function(x) as.numeric(as.character(x))))
    exposure <- exposure %>% 
      mutate(accumulation = rowSums(exp)) %>% 
      mutate(ever = ifelse(accumulation > 0, 1, 0))
    exposure$recency <-  rowSums(exp %*% diag(recency.vec))
    
    colnames(exposure)[colnames(exposure)=="accumulation"] <- "exposure"
    rm(exp)
  }
  
  else{
  colnames(exposure)[colnames(exposure) == temp$Variable_Name] <- "exposure"
  }
  exposure$sum <- rowSums(exposure[,colnames(exposure) %in% b])
  
  res <-tidy(lm(age7 ~ exposure + ed_momgest +WHITE + Female + 
                  mom_birthage + ppregnum + birthweight +
                  sustained.smoke, data = exposure))
  conf <- confint(lm(age7 ~ exposure + ed_momgest +WHITE + Female + 
                  mom_birthage + ppregnum + birthweight +
                  sustained.smoke, data = exposure))
  
  a <- data.frame(Cpg = temp$CpG,
             Variable_Name = temp$Variable_Name, 
             Model = temp$Model,
             Timing = temp$Timing,
             Adversity = temp$Adversity, 
             p.value = res$p.value[2],
             estimate = res$estimate[2],
             statistic = res$statistic[2],
             std.error = res$std.error[2],
             CI_lower = conf[2,1],
             CI_high = conf[2,2],
             beta.unexp = mean(age7[which(exposure$exposure==0)]),
             beta.exp = mean(age7[which(exposure$exposure>0)]),
             beta.unexp.all = mean(age7[which(exposure$sum==0 & exposure$exposure==0)]),
             beta.exp.other = mean(age7[which(exposure$sum>0)]))
  a$delta <- a$beta.exp - a$beta.unexp
  
  f7.associations <- rbind(f7.associations, a)
  
  rm(res, a, b, exposure, age7, temp, recency.vec)
}

head(f7.associations)
f7.associations$FDR <- p.adjust(f7.associations$p.value, method = "BH")
f7.associations$bonf <- p.adjust(f7.associations$p.value, method = "bonferroni")
sum(f7.associations$p.value<0.05) #46
sum(f7.associations$FDR<0.05) #46
sum(f7.associations$bonf<0.05) #46

save(f7.associations, file = "~/Dropbox (Partners HealthCare)/Dunn_shared/ALSPAC/Epigenetic/Paper Draft - Adversity->DNAm age17/Pacruelty_rerun/Results/Stability/age7.associations.age7.2022-12-05.Rdata")
write.table(f7.associations, file = "~/Dropbox (Partners HealthCare)/Dunn_shared/ALSPAC/Epigenetic/Paper Draft - Adversity->DNAm age17/Pacruelty_rerun/Tables/Stability/age7.associations.age7.2022-12-05.txt", sep ="\t", 
            col.names = T, row.names=F, quote=F)


#age 15
for(i in 1:nrow(new.results.fdr)){
  print(i)
  if(i == 1){f15.associations <- data.frame()}
  temp <- new.results.fdr[i,]
  age15 <- f15.beta.cor[,colnames(f15.beta.cor) == temp$CpG]
  

  b <- colnames(beast)[grep(paste0("^", temp$Adversity), colnames(beast))]
  if(temp$Adversity =="mompsych"){b <- b[-grep("wk",b)]}
  if(temp$Adversity =="oneadult"){b <- b[-grep("sum", b)]}
  recency.vec <- gsub("m", "", gsub(paste0(temp$Adversity,"_"),"",b))
  recency.vec[grep("y",recency.vec)] <- as.numeric(gsub("y","",
                                                            recency.vec[grep("y",recency.vec)]))*12
  recency.vec <- as.numeric(recency.vec)/12

  
  exposure <- beast[,colnames(beast) %in% c("cidB1471", covars, b)]
  exposure <- exposure[match(samples.f15.notwins$ALN, exposure$cidB1471),]
  age15 <- age15[complete.cases(exposure)]
  exposure <- exposure[complete.cases(exposure),]
  
  if(temp$Variable_Name == "recency"){
    exp <- do.call("cbind", lapply(exposure[, colnames(exposure) %in% b], 
                                   function(x) as.numeric(as.character(x))))
    exposure <- exposure %>% 
      mutate(accumulation = rowSums(exp)) %>% 
      mutate(ever = ifelse(accumulation > 0, 1, 0))
    exposure$recency <-  rowSums(exp %*% diag(recency.vec))
    
    colnames(exposure)[colnames(exposure)=="recency"] <- "exposure"
    rm(exp)
  }
  
  if(temp$Variable_Name == "accumulation"){
    exp <- do.call("cbind", lapply(exposure[, colnames(exposure) %in% b], 
                                   function(x) as.numeric(as.character(x))))
    exposure <- exposure %>% 
      mutate(accumulation = rowSums(exp)) %>% 
      mutate(ever = ifelse(accumulation > 0, 1, 0))
    exposure$recency <-  rowSums(exp %*% diag(recency.vec))
    
    colnames(exposure)[colnames(exposure)=="accumulation"] <- "exposure"
    rm(exp)
  }
  
  else{
  colnames(exposure)[colnames(exposure) == temp$Variable_Name] <- "exposure"
  }
  exposure$sum <- rowSums(exposure[,colnames(exposure) %in% b])
  
  res <-tidy(lm(age15 ~ exposure + ed_momgest +WHITE + Female + 
                  mom_birthage + ppregnum + birthweight +
                  sustained.smoke, data = exposure))
  conf <- confint(lm(age15 ~ exposure + ed_momgest +WHITE + Female + 
                  mom_birthage + ppregnum + birthweight +
                  sustained.smoke, data = exposure))
  
  a <- data.frame(Cpg = temp$CpG,
             Variable_Name = temp$Variable_Name, 
             Model = temp$Model,
             Timing = temp$Timing,
             Adversity = temp$Adversity, 
             p.value = res$p.value[2],
             estimate = res$estimate[2],
             statistic = res$statistic[2],
             std.error = res$std.error[2],
             CI_lower = conf[2,1],
             CI_high = conf[2,2],
             beta.unexp = mean(age15[which(exposure$exposure==0)]),
             beta.exp = mean(age15[which(exposure$exposure>0)]),
             beta.unexp.all = mean(age15[which(exposure$sum==0 & exposure$exposure==0)]),
             beta.exp.other = mean(age15[which(exposure$sum>0)]))
  a$delta <- a$beta.exp - a$beta.unexp
  
  f15.associations <- rbind(f15.associations, a)
  
  rm(res, a, b, exposure, age15, temp, recency.vec)
}

head(f15.associations)
f15.associations$FDR <- p.adjust(f15.associations$p.value, method = "BH")
f15.associations$bonf <- p.adjust(f15.associations$p.value, method = "bonferroni")
sum(f15.associations$p.value<0.05) #1
sum(f15.associations$FDR<0.05) #0
sum(f15.associations$bonf<0.05) #0
#nothing survives

save(f15.associations, file = "~/Dropbox (Partners HealthCare)/Dunn_shared/ALSPAC/Epigenetic/Paper Draft - Adversity->DNAm age17/Pacruelty_rerun/Results/Stability/age15.associations.age7.2022-12-05.Rdata")
write.table(f15.associations, file = "~/Dropbox (Partners HealthCare)/Dunn_shared/ALSPAC/Epigenetic/Paper Draft - Adversity->DNAm age17/Pacruelty_rerun/Tables/Stability/age15.associations.age7.2022-12-05.txt", sep ="\t", 
            col.names = T, row.names=F, quote=F)


rm(beast, covars, conf,
   f15.beta.cor, samples.f15.notwins,
   f15.associations, i, 
   cord.associations, cord.beta.cor, cord.samples,
   f7.associations, f7.beta.cor, f7.samples, results, new.results.fdr)

```


*Figure 3 - panel A emergence of age 15 hits*
```{r}
load("~/Dropbox (Partners HealthCare)/Dunn_shared/ALSPAC/Epigenetic/Paper Draft - Adversity->DNAm age17/Pacruelty_rerun/Results/Stability/cord.associations.age15.2022-08-24.Rdata", verbose=T)
load("~/Dropbox (Partners HealthCare)/Dunn_shared/ALSPAC/Epigenetic/Paper Draft - Adversity->DNAm age17/Pacruelty_rerun/Results/Stability/age7.associations.age15.2022-08-24.Rdata", verbose=T)
load("~/Dropbox (Partners HealthCare)/Dunn_shared/ALSPAC/Epigenetic/Paper Draft - Adversity->DNAm age17/Pacruelty_rerun/Results/Stability/age15.associations.age15.2022-08-24.Rdata", verbose=T)

plot.traj <- rbind(data.frame(CpG = rep(cord.associations$Cpg,3), 
                              Adversity = rep(cord.associations$Adversity,3),
                              Timing = rep(cord.associations$Timing, 3),
                              Model = rep(cord.associations$Model, 3),
                              Estimate = rep(cord.associations$estimate, 3),
                              Age = 0,
                              P.value = rep(ifelse(cord.associations$p.value<0.05, "p<0.05","NS"),3),
                              Group = c(rep("Unexposed", 41),
                                        rep("Exposed - SP", 41),rep("Exposed - other", 41)),
                              beta = c(cord.associations$beta.unexp.all,
                                        cord.associations$beta.exp, cord.associations$beta.exp.other)
                              ),
                   data.frame(CpG = rep(f7.associations$Cpg,3), 
                              Adversity = rep(f7.associations$Adversity,3),
                              Timing = rep(f7.associations$Timing, 3),
                              Model = rep(f7.associations$Model, 3),
                              Estimate = rep(f7.associations$estimate, 3),
                              Age = 7,
                              P.value = rep(ifelse(f7.associations$p.value<0.05, "p<0.05","NS"),3),
                              Group = c(rep("Unexposed", 41),
                                        rep("Exposed - SP", 41),rep("Exposed - other", 41)),
                              beta = c(f7.associations$beta.unexp.all,
                                        f7.associations$beta.exp, f7.associations$beta.exp.other)
                              ),
                   data.frame(CpG = rep(f15.associations$Cpg,3), 
                              Adversity = rep(f15.associations$Adversity,3),
                              Timing = rep(f15.associations$Timing, 3),
                              Model = rep(f15.associations$Model, 3),
                              Estimate = rep(f15.associations$estimate, 3),
                              Age = 15,  
                              P.value = rep(ifelse(f15.associations$p.value<0.05, "p<0.05","NS"),3),
                              Group = c(rep("Unexposed", 41),
                                        rep("Exposed - SP", 41),rep("Exposed - other", 41)),
                              beta = c(f15.associations$beta.unexp.all,
                                        f15.associations$beta.exp, f15.associations$beta.exp.other)
                              ))
dim(plot.traj)
plot.traj$Group <- factor(plot.traj$Group, 
                          levels = c("Unexposed","Exposed - other", "Exposed - SP"))

plot.traj.2 <- plot.traj[plot.traj$Group=="Unexposed" & plot.traj$Age != 0,]
dim(plot.traj.2)
plot.traj.2$Concordance <- unlist(lapply(1:nrow(plot.traj.2), function(x){
  cg <- plot.traj.2$CpG[x]
  b <- plot.traj.2$Estimate[plot.traj.2$CpG == cg]
  ifelse(b[1]<0 & b[2]<0, "Concordant", 
         ifelse(b[1]>0 & b[2]>0, "Concordant", "Non-concordant"))
}))
summary(factor(plot.traj.2$Concordance))


ggplot(plot.traj.2, aes(x = factor(Age, levels = c(7, 15)), y = Estimate, col = Concordance))+
  geom_point()+
  geom_line(aes(group = CpG), alpha=0.95)+
  theme_classic()+
  geom_hline(yintercept=0, linetype=3)+
  scale_color_manual(values = c("darkblue","grey"), labels = c("Concordant (20)",
                                                           "Non-concordant (21)"),
                     "Directionality")+
  xlab("Age at DNAm sampling")+
  ylab("Effect estimate for age 15 hits")+
  theme(text = element_text("Arial"),
        axis.text = element_text(size=12),
        axis.title = element_text(size=14),
        legend.text= element_text(size =10))+
  scale_x_discrete(expand = c(0.05, 0.05))

ggsave(file = "~/Dropbox (Partners HealthCare)/Dunn_shared/ALSPAC/Epigenetic/Paper Draft - Adversity->DNAm age17/Submission 9 - Lancet Child Adolescent Health - Resubmission 2/Figures/Figure3A_2023-04-17.eps", dpi = 300, width=8, height=5, device=cairo_ps)

rm(plot.traj, plot.traj.2, cord.associations, f7.associations, f15.associations)
```


*Figure 3 - panel D stability of age 7 hits*
```{r}
load("~/Dropbox (Partners HealthCare)/Dunn_shared/ALSPAC/Epigenetic/Paper Draft - Adversity->DNAm age17/Pacruelty_rerun/Results/Stability/age7.associations.age7.2022-08-24.Rdata", verbose=T)
load("~/Dropbox (Partners HealthCare)/Dunn_shared/ALSPAC/Epigenetic/Paper Draft - Adversity->DNAm age17/Pacruelty_rerun/Results/Stability/age15.associations.age7.2022-08-24.Rdata", verbose=T)
load("~/Dropbox (Partners HealthCare)/Dunn_shared/ALSPAC/Epigenetic/Paper Draft - Adversity->DNAm age17/Pacruelty_rerun/Results/Stability/cord.associations.age7.2022-08-24.Rdata", verbose=T)

head(cord.associations)
head(f15.associations)

plot.traj <- rbind(data.frame(CpG = rep(f7.associations$Cpg,3), 
                              Adversity = rep(f7.associations$Adversity,3),
                              Timing = rep(f7.associations$Timing, 3),
                              Model = rep(f7.associations$Model, 3),
                              Estimate = rep(f7.associations$estimate, 3),
                              Age = 7,
                              P.value = rep(ifelse(f7.associations$p.value<0.05, "p<0.05","NS"),3),
                              Group = c(rep("Unexposed", 46),
                                        rep("Exposed - SP", 46),rep("Exposed - other", 46)),
                              beta = c(f7.associations$beta.unexp.all,
                                        f7.associations$beta.exp, f7.associations$beta.exp.other)
                              ),
                   data.frame(CpG = rep(f15.associations$Cpg,3), 
                              Adversity = rep(f15.associations$Adversity,3),
                              Timing = rep(f15.associations$Timing, 3),
                              Model = rep(f15.associations$Model, 3),
                              Estimate = rep(f15.associations$estimate, 3),
                              Age = 15,  
                              P.value = rep(ifelse(f15.associations$p.value<0.05, "p<0.05","NS"),3),
                              Group = c(rep("Unexposed", 46),
                                        rep("Exposed - SP", 46),rep("Exposed - other", 46)),
                              beta = c(f15.associations$beta.unexp.all,
                                        f15.associations$beta.exp, f15.associations$beta.exp.other)
                              ))
dim(plot.traj)
plot.traj$Group <- factor(plot.traj$Group, 
                          levels = c("Unexposed","Exposed - other", "Exposed - SP"))

plot.traj.2 <- plot.traj[plot.traj$Group=="Unexposed" & plot.traj$Age != 0,]
dim(plot.traj.2)
plot.traj.2$Concordance <- unlist(lapply(1:nrow(plot.traj.2), function(x){
  cg <- plot.traj.2$CpG[x]
  b <- plot.traj.2$Estimate[plot.traj.2$CpG == cg]
  ifelse(b[1]<0 & b[2]<0, "Concordant", 
         ifelse(b[1]>0 & b[2]>0, "Concordant", "Non-concordant"))
}))
summary(factor(plot.traj.2$Concordance))/2


ggplot(plot.traj.2, aes(x = factor(Age, levels = c(7, 15)), y = Estimate, col = Concordance))+
  geom_point()+
  geom_line(aes(group = CpG), alpha = 0.95)+
  theme_classic()+
  geom_hline(yintercept=0, linetype=3)+
  scale_color_manual(values = c("darkblue","grey"), labels = c("Concordant (24)",
                                                           "Non-concordant (22)"),
                     "Directionality")+
  xlab("Age at DNAm sampling")+
  ylab("Effect estimate for age 7 hits")+
  theme(text = element_text("Arial"),
        axis.text = element_text(size=12),
        axis.title = element_text(size=14),
        legend.text= element_text(size =10))+
  scale_x_discrete(expand = c(0.05, 0.05))

ggsave(file = "~/Dropbox (Partners HealthCare)/Dunn_shared/ALSPAC/Epigenetic/Paper Draft - Adversity->DNAm age17/Submission 9 - Lancet Child Adolescent Health - Resubmission 2/Figures/Figure3B_2023-04-17.eps", dpi = 300, width=8, height=5, device=cairo_ps)

rm(plot.traj, plot.traj.2, f7.associations, f15.associations)
```
