==> 3. Analysis scripts/8. Top_hits_trajectories_2022-09-02.Rmd <==
---
title: "Part 8. Trajectories"
output: html_document
---

```{r setup}
library(dendextend)
library(cluster) 
library(broom)
library(dplyr)
library(ggplot2)
library(viridis)
library(reshape)

setwd("~/Dropbox (Partners HealthCare)/Dunn_shared/ALSPAC/Epigenetic/Paper Draft - Adversity->DNAm age17/Pacruelty_rerun/")
load("Results/age15.results.variable_2022-08-18.Rdata", verbose=T)
load("Results/age15.results.top_2022-08-24.Rdata", verbose=T)
load("~/Dropbox (Partners HealthCare)/Dunn_shared/ALSPAC/Epigenetic/Paper Draft - Adversity->DNAm age17/Pacruelty_rerun/Results/age15.data.combined_2022-08-26.Rdata")

#sample sizes by age
data.combined %>%
  group_by(CpG, Age) %>% count() %>%
  group_by(Age) %>% summarize(min = min(n), max = max(n))

adv.labels <- c("Caregiver physical or emotional abuse", 
                "Sexual or physical abuse (by anyone)", 
                "Maternal psychopathology", 
                "One adult in the household", 
                "Family instability", 
                "Financial hardship", 
                "Neighborhood disadvantage")

```

*Omnibus test*
```{r}
### OMNIBUS TEST ###

omnibus <- do.call(rbind, lapply(1:length(unique(data.combined$CpG)), function(x){
  print(x)
  cpg <- unique(data.combined$CpG)[x]
  dat <- data.combined[data.combined$CpG == cpg,]
  summary(dat$Group)
  if(sum(dat$Timing %in% c("accumulation","recency"))>0)
    {dat$Group <- factor(dat$Group, levels = c("Unexposed","Exposed-SP"))}
  data.frame(CpG = cpg,
             Timing = dat$Timing[1],
             Adversity = dat$Adversity[1],
             tidy(anova(aov(dnam~Group*Age+ID, data = dat)))[4,])
}))

head(omnibus)
sum(omnibus$p.value<0.05) #34/41 have time-varying effects
max(omnibus$p.value)
omnibus$FDR <- p.adjust(omnibus$p.value, method="BH")
sum(omnibus$FDR<0.05) #34/41
dim(omnibus[omnibus$p.value>0.05,])


omnibus.sig <- omnibus[omnibus$FDR<0.05,]

```

*Tukey HSD for subsequent cluster *
```{r}
#making names vector for the loop - messy legacy code.. 
dat <- data.combined[data.combined$CpG == "cg17928317",]
b <- do.call(rbind, TukeyHSD(aov(dnam~Group*Age, data = dat)))
dim(b)
fullhsd <- rownames(b)

#running tukey on all CpGs, but will remove CpGs that don't have group*age effects after. 
#This way, I can plot them after the fact should I want to
omni.tukey.pval <- do.call(rbind, lapply(omnibus$CpG, function(x){
  print(as.character(x))
  dat <- data.combined[data.combined$CpG == x,]
  if(sum(dat$Timing %in% c("accumulation","recency"))>0)
    {dat$Group <- factor(dat$Group, levels = c("Unexposed","Exposed-SP"))}
  
  a <- do.call(rbind, TukeyHSD(aov(dnam~Group*Age+ID, data = dat)))
  a <- a[match(fullhsd, rownames(a)),]
  rownames(a) <- fullhsd
  b <- ifelse(a[,4]<0.05, 1, 0)
  
  data.frame(CpG = x, 
               Timing = dat$Timing[1],
               Adversity = dat$Adversity[1],
               t(b))
}))

dim(omni.tukey.pval) #41, 45
head(colnames(omni.tukey.pval))
omni.tukey.pval[1:5,1:10]
colnames(omni.tukey.pval)[grep("Exposed.other.Unexposed", colnames(omni.tukey.pval))]
rownames(omni.tukey.pval) <- omni.tukey.pval$CpG
for(i in 2:ncol(omni.tukey.pval)){
  omni.tukey.pval[,i] <- factor(omni.tukey.pval[,i])
}

save(omni.tukey.pval,
     file ="~/Dropbox (Partners HealthCare)/Dunn_shared/ALSPAC/Epigenetic/Paper Draft - Adversity->DNAm age17/Pacruelty_rerun/Results/aim3.trajectories.tukeyHSD_2022-08-29.Rdata")

#focusing on CpGs that have group*age effects
omni.tukey.pval.sig <- omni.tukey.pval[omni.tukey.pval$CpG %in% omnibus.sig$CpG,]
dim(omni.tukey.pval.sig) #34/45
```


*Cluster stats function*
```{r}
library(fpc)
cstats.table <- function(dist, tree, k) {
clust.assess <- c("cluster.number","n","within.cluster.ss","average.within","average.between",
                  "wb.ratio","dunn2","avg.silwidth")
clust.size <- c("cluster.size")
stats.names <- c()
row.clust <- c()
output.stats <- matrix(ncol = k, nrow = length(clust.assess))
cluster.sizes <- matrix(ncol = k, nrow = k)
for(i in c(1:k)){
  row.clust[i] <- paste("Cluster-", i, " size")
}
for(i in c(2:k)){
  stats.names[i] <- paste("Test", i-1)
  
  for(j in seq_along(clust.assess)){
    output.stats[j, i] <- unlist(cluster.stats(d = dist, clustering = cutree(tree, k = i))[clust.assess])[j]
    
  }
  
  for(d in 1:k) {
    cluster.sizes[d, i] <- unlist(cluster.stats(d = dist, clustering = cutree(tree, k = i))[clust.size])[d]
    dim(cluster.sizes[d, i]) <- c(length(cluster.sizes[i]), 1)
    cluster.sizes[d, i]
    
  }
}
output.stats.df <- data.frame(output.stats)
cluster.sizes <- data.frame(cluster.sizes)
cluster.sizes[is.na(cluster.sizes)] <- 0
rows.all <- c(clust.assess, row.clust)
# rownames(output.stats.df) <- clust.assess
output <- rbind(output.stats.df, cluster.sizes)[ ,-1]
colnames(output) <- stats.names[2:k]
rownames(output) <- rows.all
is.num <- sapply(output, is.numeric)
output[is.num] <- lapply(output[is.num], round, 2)
output
}

```
BASED ON THIS GUIDE FOR CLUSTERING OF CATEGORICAL VALUES
https://towardsdatascience.com/hierarchical-clustering-on-categorical-data-in-r-a27e578f2995

*Clusters of interest - age, group, age-group at each age*
```{r}
interest <- c("Exposed.other.Unexposed","Exposed.SP.Unexposed","Exposed.SP.Exposed.other",
              "X7.0","X15.7",
              "Exposed.SP.0.Exposed.other.0","Exposed.SP.0.Unexposed.0",
              "Exposed.other.0.Unexposed.0","Exposed.SP.7.Exposed.other.7",
              "Exposed.SP.7.Unexposed.7","Exposed.other.7.Unexposed.7",
              "Exposed.SP.15.Exposed.other.15","Exposed.SP.15.Unexposed.15",
              "Exposed.other.15.Unexposed.15")

#----- Dissimilarity Matrix -----#
# to perform different types of hierarchical clustering
# package functions used: daisy(), diana(), clusplot()
gower.dist <- daisy(omni.tukey.pval.sig[,interest], metric = c("gower"))
# gower.dist <- daisy(omni.tukey.pval.2[omni.tukey.pval.2$CpG %in% age15.fdr$CpG,interest],
#                     metric = c("gower"))
# class(gower.dist) 
## dissimilarity , dist

#------------ DIVISIVE CLUSTERING ------------#
divisive.clust <- diana(as.matrix(gower.dist), 
                  diss = TRUE, keep.diss = TRUE)
plot(divisive.clust, main = "Divisive")

#------------ AGGLOMERATIVE CLUSTERING ------------#
aggl.clust.c <- hclust(gower.dist, method = "complete")
plot(aggl.clust.c,
     main = "Agglomerative, complete linkages")

#------------ CLUSTER STATS ------------#
stats.df.divisive <- cstats.table(gower.dist, divisive.clust, 7)
stats.df.divisive

stats.df.aggl <-cstats.table(gower.dist, aggl.clust.c, 7)
stats.df.aggl

```

*Selecting number of clusters - Figure S18*
```{r Figure S18}
# --------- Choosing the number of clusters ---------#
## Elbow
# Divisive clustering
a <- ggplot(data = data.frame(t(cstats.table(gower.dist, divisive.clust, 15))), 
  aes(x=cluster.number, y=within.cluster.ss)) + 
  geom_point()+
  geom_line()+
  theme_classic()+
  geom_vline(xintercept = 5, linetype=2, col ="darkred")+
  ggtitle("Divisive clustering") +
  labs(x = "Number of trajectory types", y = "Within trajectory type sum of squares (SS)") +
  #theme(plot.title = element_text(hjust = 0.5))+
  scale_x_continuous(breaks = seq(2,14,by=2))+
  ylim(0,0.5)

# Agglomerative clustering
b <- ggplot(data = data.frame(t(cstats.table(gower.dist, aggl.clust.c, 15))), 
  aes(x=cluster.number, y=within.cluster.ss)) + 
  geom_point()+
  geom_line()+
  theme_classic()+
  ggtitle("Agglomerative clustering") +
  labs(x = "Num.of clusters", y = "Within clusters sum of squares (SS)") +
  theme(plot.title = element_text(hjust = 0.5))+
  ylim(0,0.5)

## Silhouette
c<-ggplot(data = data.frame(t(cstats.table(gower.dist, divisive.clust, 15))), 
  aes(x=cluster.number, y=avg.silwidth)) + 
  geom_point()+
  geom_line()+
  theme_classic()+
  geom_vline(xintercept = 5, linetype=2, col ="darkred")+
  #ggtitle("Divisive clustering") +
  labs(x = "Number of trajectory types", y = "Average silhouette width") +
  #theme(plot.title = element_text(hjust = 0.5))+
  scale_x_continuous(breaks = seq(2,14, by=2))+
  ylim(0.2,.85)

d<-ggplot(data = data.frame(t(cstats.table(gower.dist, aggl.clust.c, 15))), 
  aes(x=cluster.number, y=avg.silwidth)) + 
  geom_point()+
  geom_line()+
  theme_classic()+
  #ggtitle("Agglomerative clustering") +
  labs(x = "Num.of clusters", y = "Average silhouette width") +
  theme(plot.title = element_text(hjust = 0.5))+
  ylim(0.2,.85)

## Number of 1 cluster 
t <- data.frame(t(cstats.table(gower.dist, divisive.clust, 15)))
t$oneclust <- rowSums(t[,9:23]==1)
e<- ggplot(data = t, 
  aes(x=cluster.number, y=oneclust)) + 
  geom_point()+
  geom_line()+
  theme_classic()+
  geom_vline(xintercept = 5, linetype=2, col ="darkred")+
  #ggtitle("Divisive clustering") +
  labs(x = "Number of trajectory types", y = "Number of 1 CpG trajectory types") +
  #theme(plot.title = element_text(hjust = 0.5))+
  scale_y_continuous(breaks = seq(2,14,by=2))+
  scale_x_continuous(breaks = seq(2,14,by=2))

t.2 <- data.frame(t(cstats.table(gower.dist, aggl.clust.c, 15)))
t.2$oneclust <- rowSums(t.2[,9:23]==1)
f<- ggplot(data = t.2, 
  aes(x=cluster.number, y=oneclust)) + 
  geom_point()+
  geom_line()+
  theme_classic()+
  geom_vline(xintercept = 5, linetype=2, col ="darkred")+
  #ggtitle("Agglomerative clustering") +
  labs(x = "Number of trajectory types", y = "Number of 1 CpG trajectory types") +
  theme(plot.title = element_text(hjust = 0.5))+
  scale_y_continuous(breaks = seq(2,14,by=2))+
  scale_x_continuous(breaks = seq(2,14,by=2))

grid.arrange(a,b,c,d,e,f) #agglomerative has a lot more 1 cpg trajectories, which is meh

#Figure S18 - 7 clusters is the way. 
grid.arrange(e +ggtitle("A")+theme(plot.title.position = "plot",
                                   plot.title = element_text(size=22, 
                                                             vjust = -1, face = "bold")),
             a+ggtitle("B")+theme(plot.title.position = "plot",
                                  plot.title = element_text(size=22, 
                                                             vjust = -1, face = "bold")),
             c+ggtitle("C")+theme(plot.title.position = "plot",
                                   plot.title = element_text(size=22, 
                                                             vjust = -1, face = "bold")),
             nrow=1)
rm(a,b,c,d,e,f,t, t.2)

```
Divisive clustering with 7 clusters fits the metrics best: 
- <5% of CpGs in lone clusters (2 CpGs of 41)
- lowest sum square
- max silhouette

*Dendrogram - Figure S18*
```{r Figure S18}
dendro <- as.dendrogram(divisive.clust)
dendro.col <- dendro %>%
  set("branches_k_color", k = 5, 
      value =  viridis(n=5, end = 0.9)) %>%
  set("branches_lwd", 0.6) %>%
  set("labels_colors", 
      value = c("black")) %>% 
  set("labels_cex", 0.7)
ggd1 <- as.ggdend(dendro.col)

ggplot(ggd1, theme = theme_classic(), horiz=T) +
  theme(axis.text.y = element_blank(),
        axis.line.y = element_blank(),
        axis.title.y = element_blank(),
        axis.ticks.y = element_blank())+
  labs(x = "Num. observations", y = "Height", title = "Dendrogram, k = 5")

# Radial plot looks less cluttered (and cooler), but can't get the labels fixed 
ggplot(ggd1, labels = T) +
  scale_y_reverse(expand = c(0.2, 0)) +
  coord_curvedpolar(theta="x", clip ="on")
# library(geomtextpath)
# ggplot(ggd1, labels = T) +
#   scale_y_reverse(expand = c(0.2, 0)) +
#   coord_curvedpolar(theta="x", clip ="off")
#   
```

*Saving results*
```{r}
clust.num <- cutree(divisive.clust, k = 5)
clust.num ## ADD IN THE OTHER CPGS 
traj.clusters.cl <- omni.tukey.pval[,c("CpG","Adversity", "Timing", interest)]
traj.clusters.cl$clust.num <- 0
traj.clusters.cl$clust.num[match(names(clust.num), traj.clusters.cl$CpG)] <- clust.num
head(traj.clusters.cl)
omnibus[omnibus$CpG == "cg06812747",] #sanity check, should be 0

save(traj.clusters.cl,
     file = "~/Dropbox (Partners HealthCare)/Dunn_shared/ALSPAC/Epigenetic/Paper Draft - Adversity->DNAm age17/Pacruelty_rerun/Results/aim3.FIVEtrajectories.stats_2022-09-01.Rdata")

```

*Heatmap - Figure S19*
```{r}
#heatmap

traj.clusters.long <- melt(data.frame(lapply(traj.clusters.cl, as.character),
                                      stringsAsFactors=FALSE), 
                  id = c("CpG", "clust.num","Adversity","Timing"), factorsAsStrings=T)
head(traj.clusters.long)

#Summary of CpGs / cluster / variable
traj.clusters.q <- traj.clusters.long %>%
  group_by(clust.num, variable) %>%
  mutate(count = sum(value==1, na.rm=T)) %>%
  mutate(nCpG = n_distinct(CpG)) %>%
  distinct(clust.num, variable, nCpG, count)
traj.clusters.q

# calculating the percent of each factor level in the absolute count of cluster members
traj.clusters.perc <- traj.clusters.q %>%
  group_by(clust.num, variable) %>%
  mutate(perc = count / nCpG) %>%
  arrange(clust.num)
traj.clusters.perc
traj.clusters.perc$variable <- factor(traj.clusters.perc$variable, 
                               levels = rev(interest[c(4:5,2,3,1,
                                                   7,6,8, 10,9,11,
                                                   13,12,14)]))
table(clust.num) #2,4,6,5,1
#agglomerative order - factor(clust.num, levels =c (2,4,5,3,7,6,1))
#divisive order - factor(clust.num, levels =c (3,6,7,5,4,2,1))
# ggplot(traj.clusters.perc, aes(x = factor(clust.num, levels =c (3,6,5,4,2,1)), 
#                         y = factor(variable))) +

#seven clusters - factor(clust.num, levels = c(0,1,7,3,5,2,6,4) 

#FIGURE S18
ggplot(traj.clusters.perc, aes(x = factor(clust.num, levels = c(2,4,3,5,1,0)),
                                            #c(0,1,5,3,4,2) ),
                        y = factor(variable))) +
  geom_tile(aes(fill = perc*100), alpha = 0.85, col ="black")+
  geom_hline(yintercept = 0.5) + 
  geom_hline(yintercept = 3.5) + 
  geom_hline(yintercept = 6.5) +
  geom_hline(yintercept = 9.5) +
  geom_hline(yintercept = 12.5) + 
  geom_hline(yintercept = 14.5)+
  scale_fill_gradient("% CpGs with p<0.05", low = "white", high = "#440154FF")+
  #scale_fill_gradient2(low = "darkslategray1", mid = "yellow", high = "turquoise4", 
  #                     "Fraction of trajectories")+
  scale_y_discrete(labels = rev(c(expression(bold("Age")~
                                         "                                        7 vs 0"),
                                  "15 vs 7",
                               expression(bold(Group)~"    "~Exposed[SP]~vs~Unexposed),
                               expression(Exposed[SP]~vs~Exposed[other]),
                               expression(Exposed[other]~vs~Unexposed),
                               expression(bold("Age 0")~"     "~Exposed[SP]~vs~Unexposed),
                               expression(Exposed[SP]~vs~Exposed[other]),
                               expression(Exposed[other]~vs~Unexposed),
                               expression(bold("Age 7")~"     "~Exposed[SP]~vs~Unexposed),
                               expression(Exposed[SP]~vs~Exposed[other]),
                               expression(Exposed[other]~vs~Unexposed),
                               expression(bold("Age 15")~"   "~Exposed[SP]~vs~Unexposed),
                               expression(Exposed[SP]~vs~Exposed[other]),
                               expression(Exposed[other]~vs~Unexposed)
                               ))
                              )+
  # scale_x_discrete(labels = c("Emergent (1)","Flat emergent (2)","Overcompensation (3)",
  #                             "Latent graded (4)", "Latent sensitive period (5)",
  #                             "Primed (6)"))+
  scale_x_discrete(labels = c("Latent (17)", "Overcompensation (9)", "Emergent (5)",
                              "Flat emergent (2)", "Primed (1)", "Time-stable (7)"))+
  theme_classic()+
  theme(text = element_text(family = "Arial"),
        axis.text.x = element_text(angle=45, hjust=1),
        axis.text.y = element_text(hjust= 1),
        #legend.position="top",
        axis.text = element_text(size =12),
        axis.title = element_text(size=14)
        )+
  xlab("Cluster (# of CpGs)")+
  ylab("")


clust.ref <- c("Latent", "Overcompensation", "Emergent",
                              "Flat emergent", "Primed", "Time-stable")
names(clust.ref) <- c(2,4,3,5,1,0)
```


*Analyzing the trajectories*
```{r Figures S20 and S21}
load("~/Dropbox (Partners HealthCare)/Dunn_shared/ALSPAC/Epigenetic/Paper Draft - Adversity->DNAm age17/Pacruelty_rerun/Results/aim3.trajectories.stats_2022-08-29.Rdata")

identical(as.character(traj.clusters.cl$CpG), age15.fdr.R2$CpG) #true
age15.fdr.R2$cluster <- traj.clusters.cl$clust.num
age15.fdr.R2$clust.name <- clust.ref[match(age15.fdr.R2$cluster, names(clust.ref))]
colnames(age15.fdr.R2)
write.table(age15.fdr.R2[,c("CpG","Adversity","Timing","FDR","cluster","clust.name")], 
            file = "~/Dropbox (Partners HealthCare)/Dunn_shared/ALSPAC/Epigenetic/Paper Draft - Adversity->DNAm age17/Pacruelty_rerun/Tables/aim3.five.clusters_2022-09-02.txt",
            sep="\t",quote=F, row.names=F, col.names=T)



cluster.dat <- rbind(data.frame(age15.fdr.R2, Subset = "R2-threshold"),
                     data.frame(age15.fdr.R2[age15.fdr.R2$CpG %in% age15.fdr$CpG,],
                                Subset = "FDR-significant"))
cluster.dat$cluster <- as.factor(cluster.dat$cluster)

traj.summary <- data.combined %>%
  group_by(CpG, Age, Group) %>%
  mutate(DNAm = mean(dnam)) %>%
  distinct(CpG, Adversity, Timing, Age, Group, DNAm)

traj.summary$Cluster <- age15.fdr.R2$cluster[match(traj.summary$CpG, age15.fdr.R2$CpG)]
head(traj.summary, n =18)
traj.summary$Age <- as.numeric(as.character(traj.summary$Age))

traj.summary$Adversity.2 <- adv.labels[match(traj.summary$Adversity, names(adv.labels))]
traj.summary$Adversity.2 <- factor(traj.summary$Adversity.2, 
                                   levels = adv.labels)
traj.summary$Timing.2 <- age15.fdr.R2$timeclass[match(traj.summary$CpG, age15.fdr.R2$CpG)]
traj.summary$Timing.2 <- gsub("\n"," ", traj.summary$Timing.2)

#reordering for facets
traj.summary <- traj.summary %>%
  arrange(Cluster, Adversity.2, Timing)
traj.summary$FDR <- ifelse(traj.summary$CpG %in% age15.fdr$CpG, "*","")
traj.summary$facet <- paste(paste0(traj.summary$FDR, traj.summary$CpG, traj.summary$FDR), 
                            traj.summary$Adversity.2, traj.summary$Timing.2, sep ="\n")

traj.summary$facet <- factor(traj.summary$facet, levels = unique(traj.summary$facet))
unique(traj.summary$facet)

table(age15.fdr.R2$cluster)

#Plotting trajectory plots for figure S20
plot.groups <- list(g1 = unique(traj.summary$CpG)[1:8], #clusters 0, 1
                    g2 = unique(traj.summary$CpG)[9:17], #cluster 2
                    g3 = unique(traj.summary$CpG)[18:25], #cluster 2
                    g4 = unique(traj.summary$CpG)[c(26:30, 40:41)], #cluster 3, 5
                    g5 = unique(traj.summary$CpG)[31:39]) #cluster4
traj.summary[traj.summary$CpG == unique(traj.summary$CpG)[40],]

# label.cols <- data.frame(CpG = unique(traj.summary$CpG))
# label.cols$facet_col <- ifelse(label.cols$CpG %in% age15.fdr$CpG, "grey","white")

#formatter to get same decimals
formatter <- function(...){
  function(x) format(round(x, 1), ...)
}

#Clusters 0 (7), 1 (1)

ggplot(traj.summary[traj.summary$CpG %in% plot.groups$g1,],
       aes(x= Age,
           y= DNAm*100, col=Group, group = Group))+
  geom_line(size = 1, aes(lty = Group))+
  theme_classic()+
  facet_wrap(~facet, scales = "free_y", nrow=3, ncol=3)+
  scale_color_manual(values = c("black", "blue","red"))+
  #scale_x_continuous(breaks = "c(0, 7, 15)")+
  theme(axis.title = element_blank(),
        axis.text = element_text(size=12),
        axis.text.x = element_blank(),
        axis.ticks.x = element_blank(),
        strip.text = element_text(size=9.5),
        legend.position = "none",
        legend.text = element_text(size=10),
        text = element_text(family = "Arial"))+
  ylab("Percent DNA methylation")+
  scale_y_continuous(labels = formatter(nsmall = 1))

#Cluster 2 (17)
ggplot(traj.summary[traj.summary$CpG %in% plot.groups$g2,],
       aes(x=Age, y= DNAm*100, col=Group, group = Group))+
  geom_line(size = 1, aes(lty = Group))+
  theme_classic()+
  facet_wrap(~facet, scales = "free_y", nrow=3, ncol=3)+
  scale_color_manual(values = c("black", "blue","red"))+
  scale_x_continuous(breaks = c(0, 7, 15))+
  theme(axis.title = element_blank(),
        axis.text = element_text(size=12),
        axis.text.x = element_blank(),
        axis.ticks.x = element_blank(),
        strip.text = element_text(size=9.5),
        legend.position = "none",
        legend.text = element_text(size=10),
        text = element_text(family = "Arial"))+
  ylab("Percent DNA methylation")+
  scale_y_continuous(labels = formatter(nsmall = 1))


ggplot(traj.summary[traj.summary$CpG %in% plot.groups$g3,],
       aes(x=Age, y= DNAm*100, col=Group, group = Group))+
  geom_line(size = 1, aes(lty = Group))+
  theme_classic()+
  facet_wrap(~facet, scales = "free_y", nrow=3, ncol=3)+
  scale_color_manual(values = c("black", "blue","red"))+
  scale_x_continuous(breaks = c(0, 7, 15))+
  theme(axis.title = element_blank(),
        axis.text = element_text(size=12),
        axis.text.x = element_blank(),
        axis.ticks.x = element_blank(),
        strip.text = element_text(size=9.5),
        legend.position = "none",
        legend.text = element_text(size=10),
        text = element_text(family = "Arial"))+
  ylab("Percent DNA methylation")+
  scale_y_continuous(labels = formatter(nsmall = 1))


#Cluster 3 (5), 5 (2)
ggplot(traj.summary[traj.summary$CpG %in% plot.groups$g4,],
       aes(x=Age, y= DNAm*100, col=Group, group = Group))+
  geom_line(size = 1, aes(lty = Group))+
  theme_classic()+
  facet_wrap(~facet, scales = "free_y", nrow=3, ncol=3)+
  scale_color_manual(values = c("black", "blue","red"))+
  scale_x_continuous(breaks = c(0, 7, 15))+
  theme(axis.title = element_blank(),
        axis.text = element_text(size=12),
        axis.text.x = element_blank(),
        axis.ticks.x = element_blank(),
        strip.text = element_text(size=9.5),
        legend.position = "none",
        legend.text = element_text(size=10),
        text = element_text(family = "Arial"))+
  ylab("Percent DNA methylation")+
  scale_y_continuous(labels = formatter(nsmall = 1))

#Cluster 4 (9)
ggplot(traj.summary[traj.summary$CpG %in% plot.groups$g5,],
       aes(x=Age, y= DNAm*100, col=Group, group = Group))+
  geom_line(size = 1, aes(lty = Group))+
  theme_classic()+
  facet_wrap(~facet, scales = "free_y", nrow=3, ncol=3)+
  scale_color_manual(values = c("black", "blue","red"))+
  scale_x_continuous(breaks = c(0, 7, 15))+
  theme(axis.title = element_blank(),
        axis.text = element_text(size=12),
        axis.text.x = element_blank(),
        axis.ticks.x = element_blank(),
        strip.text = element_text(size=9.5),
        legend.position = "none",
        legend.text = element_text(size=10),
        text = element_text(family = "Arial"))+
  ylab("Percent DNA methylation")+
  scale_y_continuous(labels = formatter(nsmall = 1))


leg <- ggplot(traj.summary,
       aes(x=Age, y= DNAm*100, col=Group, group = Group))+
  geom_line(size = 1, aes(lty = Group))+
  theme_classic()+
  scale_color_manual(values = c("black", "blue","red"))+
  theme(legend.text = element_text(size=12),
        legend.title = element_text(size=14))

grid.arrange(cowplot::get_legend(leg))

table(cluster.dat$cluster[cluster.dat$Subset =="FDR-significant"])
table(cluster.dat$cluster[cluster.dat$Subset =="R2-threshold"])
cluster.dat$cluster <- factor(cluster.dat$cluster, levels = c(2,4,3,5,1,0))


### Figure S21 -  showing clusters by R2 and FDR 
cluster.dat.2 <- cluster.dat %>%
  group_by(cluster, Subset) %>%
  count()
cluster.dat.2$freq <- ifelse(cluster.dat.2$Subset =="FDR-significant", cluster.dat.2$n/22, cluster.dat.2$n/41)
cluster.dat.2$total <- ifelse(cluster.dat.2$Subset =="FDR-significant",22 , 41)
cluster.dat.2$Subset <- factor(cluster.dat.2$Subset, levels = c("FDR-significant","R2-threshold"))


#Figure S20 - comparing FDR to R2
ggplot(cluster.dat.2, aes(x=cluster, y=freq*100, fill = Subset, group = Subset))+
  geom_bar(position="dodge", alpha = 0.7, col = "black", stat='identity'
           )+
           #aes(y=..count../sum(..count..)*100))+
  theme_classic()+
  scale_fill_manual(values = viridis(n=4), drop=F,"",
                    labels = c("FDR-significant (22 CpGs)",
                               "R2-threshold (41 CpGs)"))+
  ylab("Percentage of total CpGs")+
  scale_x_discrete("Cluster", labels = c("Latent", "Overcompensation", "Emergent",
                              "Flat emergent", "Primed", "Time-stable"))+
  theme(axis.text.x = element_text(size=12, angle=45, hjust=1),
        axis.title.x = element_text(size=14),
        axis.text.y = element_text(size=12),
        axis.title.y = element_text(size=14),
        legend.text = element_text(size=12))+
  geom_text(aes(label = n), position = position_dodge(w=0.9), vjust= -1, size =5)+
  ylim(0, 55)
  #theme(legend.position ="none")


chisq.test(data.frame(cluster.dat.2$n[cluster.dat.2$Subset =="FDR-significant"],
                      cluster.dat.2$n[cluster.dat.2$Subset =="R2-threshold"]))
#X-squared = 1.9199, df = 5, p-value = 0.8601
```

*For Table 2 (now figure 3)*
```{r}
load("~/Dropbox (Partners HealthCare)/Dunn_shared/ALSPAC/Epigenetic/Paper Draft - Adversity->DNAm age17/Pacruelty_rerun/Results/Stability/cord.associations.age15.2022-12-07.Rdata", verbose=T)
load("~/Dropbox (Partners HealthCare)/Dunn_shared/ALSPAC/Epigenetic/Paper Draft - Adversity->DNAm age17/Pacruelty_rerun/Results/Stability/age7.associations.age15.2022-12-07.Rdata", verbose=T)
load("~/Dropbox (Partners HealthCare)/Dunn_shared/ALSPAC/Epigenetic/Paper Draft - Adversity->DNAm age17/Pacruelty_rerun/Results/Stability/age15.associations.age15.2022-12-07.Rdata", verbose=T)


plot.traj <- rbind(data.frame(CpG = rep(cord.associations$Cpg,3), 
                              Adversity = rep(cord.associations$Adversity,3),
                              Timing = rep(cord.associations$Timing, 3),
                              Model = rep(cord.associations$Model, 3),
                              Estimate = rep(cord.associations$estimate, 3),
                              Age = 0,
                              P.value = rep(ifelse(cord.associations$p.value<0.05, "p<0.05","NS"),3),
                              Group = c(rep("Unexposed", 41),
                                        rep("Exposed - SP", 41),rep("Exposed - other", 41)),
                              beta = c(cord.associations$beta.unexp.all,
                                        cord.associations$beta.exp, cord.associations$beta.exp.other)
                              ),
                   data.frame(CpG = rep(f7.associations$Cpg,3), 
                              Adversity = rep(f7.associations$Adversity,3),
                              Timing = rep(f7.associations$Timing, 3),
                              Model = rep(f7.associations$Model, 3),
                              Estimate = rep(f7.associations$estimate, 3),
                              Age = 7,
                              P.value = rep(ifelse(f7.associations$p.value<0.05, "p<0.05","NS"),3),
                              Group = c(rep("Unexposed", 41),
                                        rep("Exposed - SP", 41),rep("Exposed - other", 41)),
                              beta = c(f7.associations$beta.unexp.all,
                                        f7.associations$beta.exp, f7.associations$beta.exp.other)
                              ),
                   data.frame(CpG = rep(f15.associations$Cpg,3), 
                              Adversity = rep(f15.associations$Adversity,3),
                              Timing = rep(f15.associations$Timing, 3),
                              Model = rep(f15.associations$Model, 3),
                              Estimate = rep(f15.associations$estimate, 3),
                              Age = 15,  
                              P.value = rep(ifelse(f15.associations$p.value<0.05, "p<0.05","NS"),3),
                              Group = c(rep("Unexposed", 41),
                                        rep("Exposed - SP", 41),rep("Exposed - other", 41)),
                              beta = c(f15.associations$beta.unexp.all,
                                        f15.associations$beta.exp, f15.associations$beta.exp.other)
                              ))
dim(plot.traj)
plot.traj$Group <- factor(plot.traj$Group, 
                          levels = c("Unexposed","Exposed - other", "Exposed - SP"))

plot.traj$Adversity <- factor(plot.traj$Adversity, 
                              levels = c("parcruelty", "abuse", "mompsych",
                                         "oneadult","r_faminstability", 
                                         "Fscore","nbhqual"))

adv.labels <- c("Caregiver physical/emotional abuse", 
                "Sexual/physical abuse", 
                "Maternal psychopathology", 
                "One adult in the household", 
                "Family instability", 
                "Financial hardship", 
                "Neighborhood disadvantage")

plot.traj$Adversity.2 <- ifelse(plot.traj$Adversity == "parcruelty", adv.labels[1],
                                ifelse(plot.traj$Adversity == "abuse", adv.labels[2],
                                       ifelse(plot.traj$Adversity == "mompsych", adv.labels[3],
                                              ifelse(plot.traj$Adversity == "oneadult",
                                                     adv.labels[4],
                                                     ifelse(plot.traj$Adversity == "Fscore",
                                                            adv.labels[6], plot.traj$Adversity
                                                            )))))
plot.traj$Adversity.2 <- factor(plot.traj$Adversity.2, levels = adv.labels)

plot.traj$Timing.2 <- ifelse(plot.traj$Timing == "very_early","Very early childhood", 
                             ifelse(plot.traj$Timing == "early","Early childhood", 
                                    ifelse(plot.traj$Timing == "middle","Middle childhood", 
                                           ifelse(plot.traj$Timing == "late","Late childhood",
                                                  "Accumulation"
                                           ))))


plot.traj$facets <- paste(plot.traj$Adversity.2, plot.traj$Timing.2, plot.traj$CpG, sep="\n")
plot.traj$facets <- factor(plot.traj$facets, levels = unique(plot.traj$facets))

ggplot(plot.traj[,],
       aes(x = Age, y= beta*100, col = Group, group = Group))+
  geom_point()+
  geom_line()+
  theme_classic()+
  facet_wrap(~facets, scales="free")+
  scale_color_manual(values = c("black", "blue","red"))+
  scale_x_continuous(breaks = c(0, 7, 15))+
  theme(axis.title = element_text(size = 14),
        text = element_text(family = "Arial"))+
  ylab("Percent DNA methylation")

plot.traj.specific <- plot.traj[plot.traj$CpG %in% 
                                  c("cg26970800","cg16907527","cg05491478", "cg15454534",
                                    #"cg04036644",
                                    "cg17928317","cg16813552"), ]
plot.traj.specific

plot.traj.specific$Cluster <- "None"
plot.traj.specific$Cluster[plot.traj.specific$CpG == "cg26970800"] <- "Emergent"
plot.traj.specific$Cluster[plot.traj.specific$CpG == "cg16907527"] <- "Flat emergent"
plot.traj.specific$Cluster[plot.traj.specific$CpG == "cg05491478"] <- "Overcompensation"
plot.traj.specific$Cluster[plot.traj.specific$CpG == "cg15454534"] <- "Latent"
#plot.traj.specific$Cluster[plot.traj.specific$CpG == "cg04036644"] <- "Latent sensitive period"
plot.traj.specific$Cluster[plot.traj.specific$CpG == "cg17928317"] <- "Primed"
plot.traj.specific$Cluster[plot.traj.specific$CpG == "cg16813552"] <- "Stable"

ggplot(plot.traj.specific[,],
       aes(x = Age, y= beta*100, col = Group, group = Group))+
  geom_point(show_guide=FALSE)+
  geom_line(size = 1, aes(lty = Group))+
  theme_light()+
  facet_wrap(~Cluster, scales="free", ncol=3)+
  scale_color_manual(values = c("black", "blue","red"))+
  scale_linetype_manual(values = c(1,3,2))+
  scale_x_continuous(breaks = c(0, 7, 15))+
  theme(axis.title = element_text(size = 14),
        axis.text = element_text(size=12),
        axis.text.y = element_blank(),
        axis.ticks.y = element_blank(),
        strip.text = element_text(color="black", size =12),
        legend.text = element_text(size=10),
        strip.background = element_rect(fill = NA, color=NA),
        panel.border = element_rect(colour = "black", fill = NA),
        #legend.position = c(0.065,0.7),
        legend.background = element_blank()
        )+
  ylab("DNA methylation levels")+
  xlab("Age at DNA methylation collection (years)")

ggsave(file = "~/Dropbox (Partners HealthCare)/Dunn_shared/ALSPAC/Epigenetic/Paper Draft - Adversity->DNAm age17/Submission 9 - Lancet Child Adolescent Health - Resubmission 2/Figures/Figure4_2023-04-17.eps", dpi = 300, width=10, height=5, device=cairo_ps)

```
