==> 3. Analysis scripts/9. Additional_analyses_2022-12-09.Rmd <==
---
title: "Untitled"
output: html_document
---

```{r}
dim(age15.results.variable[age15.results.variable$R2>0.035,])
max(age15.fdr.R2$P.value) #5.94x10-6
max(age15.fdr.R2$FDR)

```

*QQ plots*
```{r}
load(file="/Users/alexlussier/Dropbox (Partners HealthCare)/Dunn_shared/ALSPAC/Epigenetic/Paper Draft - Adversity->DNAm age17/Pacruelty_rerun/Results/age15.results.variable_2022-08-18.Rdata", verbose=T)


gg_qqplot <- function(ps, ci = 0.95) {
  n  <- length(ps)
  df <- data.frame(
    observed = -log10(sort(ps)),
    expected = -log10(ppoints(n)),
    clower   = -log10(qbeta(p = (1 - ci) / 2, shape1 = 1:n, shape2 = n:1)),
    cupper   = -log10(qbeta(p = (1 + ci) / 2, shape1 = 1:n, shape2 = n:1))
  )
  log10Pe <- expression(paste("Expected -log"[10], plain(P)))
  log10Po <- expression(paste("Observed -log"[10], plain(P)))
  ggplot(df) +
    geom_ribbon(
      mapping = aes(x = expected, ymin = clower, ymax = cupper),
      alpha = 0.1
    ) +
    geom_point(aes(expected, observed), shape = 1, size = 3) +
    geom_abline(intercept = 0, slope = 1, alpha = 0.5) +
    # geom_line(aes(expected, cupper), linetype = 2, size = 0.5) +
    # geom_line(aes(expected, clower), linetype = 2, size = 0.5) +
    xlab(log10Pe) +
    ylab(log10Po)
}

inflation <- function(ps) {
  chisq <- qchisq(1 - ps, 1)
  lambda <- median(chisq) / qchisq(0.5, 1)
  lambda
}

adv.label <- unique(age15.results.variable$Adversity)
names(adv.label) <- c("Sexual or physical abuse (by anyone)","Financial hardship", 
                      "Family instability", "Maternal psychopathology", "Neighborhood disadvantage",
                      "One adult in the household", "Caregiver physical or emotional abuse")
             
setwd("/Users/alexlussier/Dropbox (Partners HealthCare)/Dunn_shared/ALSPAC/Epigenetic/Paper Draft - Adversity->DNAm age17/Submission 8 - Lancet Child Adolescent Health - Resubmission/Results/Figures/qqplots")         
for(i in adv.label){
  print(i)
  ps <- age15.results.variable$P.value[age15.results.variable$Adversity==i]
  p <- gg_qqplot(ps) +
    theme_bw(base_size = 20) +
    annotate(
      geom = "text",
      x = -Inf,
      y = Inf,
      hjust = -0.15,
      vjust = 1 + 0.15 * 3,
      label = sprintf("λ = %.2f", inflation(ps)),
      size = 8
    ) +
    ggtitle(names(adv.label[adv.label==i]))+
    theme(
      axis.ticks = element_line(size = 0.5),
      panel.grid = element_blank()
      # panel.grid = element_line(size = 0.5, color = "grey80")
      )
  png(paste0(i, "_qq_",Sys.Date(), ".png"))
  print(p)
  dev.off()
}


rm(age15.results.variable)
```


*Stability over time*
```{r}

load("~/Dropbox (Partners HealthCare)/Dunn_shared/ALSPAC/Epigenetic/Paper Draft - Adversity->DNAm age17/Pacruelty_rerun/Results/age15.data.combined_2022-08-26.Rdata")

library(tidyr)
for(i in unique(data.combined$CpG)){
  if(i == unique(data.combined$CpG)[1]){stab <- data.frame()}
  print(i)
  dat <- data.combined[data.combined$CpG ==i,]
  dat0 <- dat[dat$Age ==0,]
  dat7 <- dat[dat$Age ==7,]
  dat15 <- dat[dat$Age ==15,]
  
  dat0 <- dat0[match(dat15$ID, dat0$ID),]
  dat7 <- dat7[match(dat15$ID, dat7$ID),]
  cor(dat0$dnam, dat15$dnam, use = "pairwise.complete.obs", method = "pearson")
  
  # cor(data.frame(age0 = dat0$dnam, 
  #                age7 = dat7$dnam, 
  #                age15= dat15$dnam), use = "pairwise.complete.obs")
  temp <- data.frame(CpG = i, 
                     Adversity = dat$Adversity[1],
                     Timing = dat$Timing[1],
                     age0_7 = cor(dat0$dnam,  dat7$dnam, use= "pairwise.complete.obs" ),
                     age0_15 = cor(dat0$dnam,  dat15$dnam, use= "pairwise.complete.obs" ),
                     age7_15 = cor(dat15$dnam,  dat7$dnam, use= "pairwise.complete.obs" ))
  stab <- rbind(stab, temp)
  
  rm(dat, dat0, dat7, dat15, temp)
  }

head(stab)
stab.melt <- reshape::melt(stab, id.vars = c("CpG","Adversity","Timing"))

stab.melt$unique <- paste(stab.melt$CpG, stab.melt$Adversity.2, 
                                       stab.melt$Timing.2, sep= " | ")

library(ggplot2)
a <- ggplot(stab.melt, aes(x= variable, y  = CpG, fill = value))+
  geom_tile()+
  theme_classic()+
  scale_fill_gradient2(low = "blue", high = "red", mid = "white", 
   midpoint = 0, limit = c(-1,1), space = "Lab", 
   name="Pearson\nCorrelation")+
  scale_x_discrete(labels = c("Age 0 - Age 7", "Age 0 - Age 15", "Age 7 - Age 15"), 
                   "Correlation between ages")+
  geom_text(data = stab.melt[abs(stab.melt$value)>0.2,], aes(label = round(value, 2)))+
  ylab("")+
  theme(axis.text = element_text(size =10),
        axis.title = element_text(size =12),
        legend.text = element_text(size =10),
        legend.title = element_text(size =12))


b <- ggplot(stability, aes(y = CpG, x= mean*100, col = as.factor(Age)))+
  geom_point(position=position_dodge(w= 0.8))+
  geom_errorbar(aes(xmin = (mean-sd)*100, xmax = (mean+sd)*100), width=0.3, position=position_dodge(w=0.8))+
  theme_classic()+
  ylab("")+
  xlab("Percent DNAm ± standard deviation")+
  scale_color_viridis("Age at DNAm\ncollection",end = 0.8, discrete=T,
                      labels = c("Age 0","Age 7","Age 15"))+
  theme(axis.text = element_text(size =10),
        axis.title = element_text(size =12),
        legend.text = element_text(size =10),
        legend.title = element_text(size =12))+
  xlim(0,100)

grid.arrange(b,a + theme(axis.text.y = element_blank(),
                         axis.ticks.y = element_blank(),
                         axis.line.y= element_blank()), ncol=2)
mean(stab.melt$value)

#doing that a bit differently - doesn't work
# for(i in unique(data.combined$CpG)){
#   if(i == unique(data.combined$CpG)[1]){stability <- data.frame()}
#   print(i)
#   dat <- data.combined[data.combined$CpG ==i,]
#   dat0 <- dat[dat$Age ==0,]
#   dat7 <- dat[dat$Age ==7,]
#   dat15 <- dat[dat$Age ==15,]
#   
#   dat0 <- dat0[match(dat15$ID, dat0$ID),]
#   dat7 <- dat7[match(dat15$ID, dat7$ID),]
#   
#   # cor(data.frame(age0 = dat0$dnam, 
#   #                age7 = dat7$dnam, 
#   #                age15= dat15$dnam), use = "pairwise.complete.obs")
#   
#   age0 <- data.frame(CpG = i, 
#                      Age =0,
#                      N= sum(!is.na(dat0$ID)),
#                      mean= mean(dat0$dnam, na.rm = T),
#                      sd = sd(dat0$dnam, na.rm = T), 
#                      cor0 = 1, 
#                      cor7 = cor(dat0$dnam,  dat7$dnam, use= "pairwise.complete.obs" ),
#                      cor15 = cor(dat0$dnam,  dat15$dnam, use= "pairwise.complete.obs" ))
#   
#   age7 <- data.frame(CpG = i, 
#                      Age =7,
#                      N= sum(!is.na(dat7$ID)),
#                      mean= mean(dat7$dnam, na.rm = T),
#                      sd = sd(dat7$dnam, na.rm = T), 
#                      cor0 = cor(dat0$dnam,  dat7$dnam, use= "pairwise.complete.obs" ),
#                      cor7 = 1,
#                      cor15 = cor(dat7$dnam,  dat15$dnam, use= "pairwise.complete.obs" ))
#   
#   age15 <- data.frame(CpG = i, 
#                       Age = 15,
#                       N= sum(!is.na(dat15$ID)),
#                       mean= mean(dat15$dnam, na.rm = T),
#                       sd = sd(dat15$dnam, na.rm = T), 
#                       cor0 = cor(dat0$dnam,  dat15$dnam, use= "pairwise.complete.obs" ), 
#                       cor7 = cor(dat7$dnam,  dat15$dnam, use= "pairwise.complete.obs" ),
#                       cor15 = 1)
#   
#   
#   temp <- rbind(age0, age7, age15)
#   stability <- rbind(stability, temp)
#   
#   rm(dat, dat0, dat7, dat15, age0, age7, age15, temp)
#   }
# 
# head(stability)
# stability.melt <- reshape::melt(stability[,-3], id.vars = c("CpG","Age"))
# head(stability.melt)
# stability.melt$variable <- factor(stability.melt$variable, levels = unique(stability.melt$variable))
# head(stability.melt)
# stability.melt$plot <- paste(stability.melt$CpG, stability.melt$Age)
# stability.melt$plot <- factor(stability.melt$plot, levels = rev(unique(stability.melt$plot)))
# stability.melt$value[stability.melt$value ==1] <- NA
# 
# #doesn't work
# ggplot(stability.melt[grep("cor",stability.melt$variable),], 
#        aes(y= plot, x = variable, fill =value))+
#   geom_tile()+
#   theme_classic()+
#   ylab("")
# 

  

```

*EWAS catalog lookup*
```{r}
load("~/Dropbox (Partners HealthCare)/Dunn_shared/ALSPAC/Epigenetic/Paper Draft - Adversity->DNAm age17/Pacruelty_rerun/Results/age15.results.top_2022-08-24.Rdata",
     verbose=T)

catalog <- data.table::fread("~/Dropbox (Partners HealthCare)/Dunn_shared/ALSPAC/Epigenetic/Aim 3/Annotations/EWAScatalog/ewascatalog-results.txt")
studies <- data.table::fread("~/Dropbox (Partners HealthCare)/Dunn_shared/ALSPAC/Epigenetic/Aim 3/Annotations/EWAScatalog/ewascatalog-studies.txt")


head(catalog)
catalogHits <- catalog[catalog$CpG %in% age15.fdr.R2$CpG, ]

dim(catalogHits) #127
length(unique(catalogHits$CpG)) #41
sort(table(catalogHits$CpG))
catalogHits[catalogHits$CpG =="cg00807464",]

#annotations
catalogHits <- cbind(catalogHits, 
                    studies[match(catalogHits$StudyID, studies$StudyID), ])
catalogHits$FDR <- ifelse(catalogHits$CpG %in% age15.fdr$CpG, "FDR","R2")
catalogHits$Adversity <- age15.fdr.R2$Adversity[match(catalogHits$CpG, age15.fdr.R2$CpG)]
catalogHits$Timing <- age15.fdr.R2$Timing[match(catalogHits$CpG, age15.fdr.R2$CpG)]

write.table(catalogHits, file = "~/Dropbox (Partners HealthCare)/Dunn_shared/ALSPAC/Epigenetic/Paper Draft - Adversity->DNAm age17/Submission 8 - Lancet Child Adolescent Health - Resubmission/Results/EWAScatalogHits_2022-12-02.txt",
            sep ="\t",quote=F, col.names=T, row.names=F)

```

*missMethyl Gene ontology*
```{r}
##missmethyl Gene ontology
library(missMethyl)
load(file="/Users/alexlussier/Dropbox (Partners HealthCare)/Dunn_shared/ALSPAC/Epigenetic/Paper Draft - Adversity->DNAm age17/Pacruelty_rerun/Results/age15.results.variable_2022-08-18.Rdata", verbose=T)

load("~/Dropbox (Partners HealthCare)/Dunn_shared/ALSPAC/Epigenetic/Paper Draft - Adversity->DNAm age17/Pacruelty_rerun/Results/age15.results.top_2022-08-24.Rdata",
     verbose=T)

library(IlluminaHumanMethylation450kanno.ilmn12.hg19)

gst.kegg <- gometh(sig.cpg = as.character(age15.fdr.R2$CpG), 
                   all.cpg = unique(age15.results.variable$CpG), collection = "KEGG")
min(gst.kegg$P.DE)
gst.kegg<- gst.kegg[order(gst.kegg$P.DE),]
head(gst.kegg)
dim(gst.kegg[gst.kegg$P.DE<0.05,])


gst.go <- gometh(as.character(age15.fdr.R2$CpG), 
                 all.cpg = unique(age15.results.variable$CpG), 
                 collection = "GO")
gst.go <- gst.go[order(gst.go$P.DE),]
gst.go[gst.go$P.DE<0.005 & gst.go$ONTOLOGY =="BP",]

head(gst.go)
gst.go[gst.go$ONTOLOGY=="CC",][1:10,]
min(gst.go$P.DE)
gst.go[1:10,]

topGO <- rbind(data.frame(Ontology = "KEGG",
                          Name = gst.kegg$Description[1:10],
                          N = gst.kegg$N[1:10],
                          DE = gst.kegg$DE[1:10],
                          P.DE = gst.kegg$P.DE[1:10]), 
               data.frame(Ontology = "Biological process",
                          Name = gst.go$TERM[gst.go$ONTOLOGY=="BP"][1:10],
                          N = gst.go$N[gst.go$ONTOLOGY=="BP"][1:10],
                          DE = gst.go$DE[gst.go$ONTOLOGY=="BP"][1:10],
                          P.DE = gst.go$P.DE[gst.go$ONTOLOGY=="BP"][1:10]),
               data.frame(Ontology = "Cellular component",
                          Name = gst.go$TERM[gst.go$ONTOLOGY=="CC"][1:10],
                          N = gst.go$N[gst.go$ONTOLOGY=="CC"][1:10],
                          DE = gst.go$DE[gst.go$ONTOLOGY=="CC"][1:10],
                          P.DE = gst.go$P.DE[gst.go$ONTOLOGY=="CC"][1:10]),
               data.frame(Ontology = "Molecular function",
                          Name = gst.go$TERM[gst.go$ONTOLOGY=="MF"][1:10],
                          N = gst.go$N[gst.go$ONTOLOGY=="MF"][1:10],
                          DE = gst.go$DE[gst.go$ONTOLOGY=="MF"][1:10],
                          P.DE = gst.go$P.DE[gst.go$ONTOLOGY=="MF"][1:10])
)

head(topGO)
topGO$Name <- factor(topGO$Name, levels = rev(topGO$Name))

ggplot(topGO, aes(y = Name, x = -log10(P.DE)))+
  geom_point()+
  geom_bar(stat='identity', width=0.1)+
  theme_classic()+
  facet_wrap(~Ontology, scales = "free_y")+
  geom_vline(xintercept = -log10(0.05), col="red",linetype=3)+
  xlab("-log10(p-value)")+
  ylab("")

```
